### 2024-08-20 97b193e
Added Contributing.md in doc folder
