+++ doc/CONTRIBUTING.md
@@ -1,0 +1,57 @@
+# Contributing to A MATLAB-based Instrument Control (MIC) package for fluorescence imaging
+
+First off, thanks for taking the time to contribute! 🚀
+
+The following is a set of guidelines for contributing to A MATLAB-based Instrument Control (MIC) package for fluorescence imaging. These are mostly guidelines, not rules. Use your best judgment, and feel free to propose changes to this document in a pull request.
+
+## How Can I Contribute?
+
+### Reporting Bugs
+
+If you've found a bug in the project, please submit an issue using the following format:
+- **Title**: A clear and concise description of the bug.
+- **Description**: Steps to reproduce, expected behavior, actual behavior, and any relevant screenshots.
+- **Environment**: Include details such as browser version, operating system, or other environment information that could be useful.
+
+### Suggesting Enhancements
+
+We welcome ideas for enhancements. Please follow this structure for enhancement suggestions:
+- **Title**: A concise description of the enhancement.
+- **Proposal**: Explain what you'd like to change or improve.
+- **Why**: Describe why this change would be beneficial.
+- **Implementation details**: If applicable, suggest how the enhancement can be implemented.
+
+### Submitting Pull Requests
+
+Before submitting a pull request, please ensure the following:
+1. Fork the repository and clone it locally.
+2. Make your changes in a new branch based on `main` (e.g., `feature/your-feature` or `fix/your-fix`).
+3. Follow the coding style of the project (run linters or formatting tools if specified).
+4. Test your changes thoroughly.
+5. Ensure that your commit messages are descriptive and meaningful.
+
+After you're done:
+1. Submit a pull request.
+2. Reference any related issues in the PR description.
+3. Await feedback or approval from maintainers.
+
+### Coding Standards
+
+- Follow the [style guide or linting rules] used in the project.
+- Write clear, concise code and comments where necessary.
+- Keep changes focused; avoid including multiple fixes or features in one pull request.
+
+### Documentation
+
+If your change affects the functionality of the project or introduces new features, please update or add to the documentation accordingly. This includes README updates and any additional documentation files if necessary.
+
+## Code of Conduct
+
+Please note that we follow a Code of Conduct to ensure a safe and welcoming environment for everyone. Make sure to review it before engaging with the community.
+
+## Additional Resources
+
+- [Project Issue Tracker](https://github.com/your-repository/issues)
+- [Code of Conduct](CODE_OF_CONDUCT.md)
+
+Thank you for contributing to A MATLAB-based Instrument Control (MIC) package for fluorescence imaging!
