+++ paper/paper.bib
@@ -141,7 +141,7 @@ number = {1},
 pages = {113603},
 year = {2024},
 issn = {2211-1247},
-doi = {https://doi.org/10.1016/j.celrep.2023.113603},
+doi = {10.1016/j.celrep.2023.113603},
 url = {https://www.sciencedirect.com/science/article/pii/S2211124723016157},
 author = {Krishna C. Mudumbi and Eric A. Burns and David J. Schodt and Zaritza O. Petrova and Anatoly Kiyatkin and Lucy W. Kim and Emma M. Mangiacapre and Irais Ortiz-Caraveo and Hector {Rivera Ortiz} and Chun Hu and Kumar D. Ashtekar and Keith A. Lidke and Diane S. Lidke and Mark A. Lemmon},
 keywords = {epidermal growth factor receptor, receptor tyrosine kinase, dimerization, single particle tracking, clustering, tyrosine kinase, transmembrane signaling, fluorescence, cancer, imaging},
