+++ paper/paper.bib
@@ -187,8 +187,9 @@ The epidermal growth factor receptor (EGFR) is a receptor tyrosine kinase with i
    TITLE = {{The PYthon Microscopy Environment}},
    HOWPUBLISHED = {https://www.python-microscopy.org/},
    YEAR = 2020,
-   NOTE = {Accessed: 2024-07-26},
+   doi = {10.5281/zenodo.4289803},
    URL = {https://www.python-microscopy.org/},
+   NOTE = {Accessed: 2024-07-26},
    ABSTRACT = {The PYthon Microscopy Environment is an open-source package providing image acquisition and data analysis functionality for a number of microscopy applications, but with a particular emphasis on super-resolution techniques (PALM/STORM/PAINT etc ...). The package is multi platform, running on Windows, Linux, and OSX.}
 }
 
