### 2024-08-21 1c9b20f
Added links to teh repo and CONTRIBUTING in the main README.
