+++ README.md
@@ -53,8 +53,6 @@ The MIC software package uses object-oriented programming where a class defines 
 This class structure is integral to the functioning and expansion of our imaging capabilities, facilitating easy maintenance and upgrading of the imaging system components.
 
 
-See [MIC Classes](doc/MICclasses.md) for the complete detailed list.
-
 ## Common Features
 Each of the instrument components in MIC have export methods, unit tests and graphical user interfaces with a common format.
 
@@ -94,3 +92,9 @@ MIC_DynamixelServo.gui
 Each instrument will be controlled by its own drivers, which must be installed on the system. In many cases, the manufacturer's software development kit (SDK) is provided to create custom applications for controlling the instrument. When installing the drivers, either a header file or dynamic-link library is installed. For example, the `MIC_MCLNanoDrive` class controls the Mad City Labs 3D Piezo stage and requires the `madlib.h` header file. During the first initialization of this class on a system, users are prompted to set the location of the `madlib.h` header file, typically located in `C:\Program Files\Mad City Labs\NanoDrive`. 
 
 Similarly, the `MIC_MCLMicroDrive` class controls the Mad City Labs Micro Stage and requires the `MicroDrive.dll` dynamic-link library. The first time this class is used on a given computer, the user will be prompted to select the location of `MicroDrive.dll`. On a Windows machine, this is typically placed by default in `C:\Program Files\Mad City Labs\MicroDrive\MicroDrive.dll` during the installation process (installation files provided by MCL).
+
+## Documentation
+The detailed documentation of each MIC class can be found here [MIC Classes](doc/MICclasses.md).
+
+## Contribution
+We welcome contributions to the MIC project. Please see the [CONTRIBUTING.md](doc/CONTRIBUTING.md) file for more information.
