### 2024-08-22 7f6e350
Added matlab-tests.yml for auto testing of abstract classes
