+++ .github/workflows/matlab-tests.yml
@@ -23,9 +23,7 @@ jobs:
 
     # Step 2: Set up MATLAB
     - name: Set up MATLAB
-      uses: matlab-actions/setup-matlab@v1
-      # with:
-      #   release: R2020a  # You can adjust this to the version you use
+      uses: matlab-actions/setup-matlab@v2
 
     # Step 3: Run the MATLAB test script
     - name: Run MATLAB tests
