### 2024-08-22 #7121 @sajjad88
@Kevin-Mattheus-Moerman thank you for helping in the initial steps. We have fixed the DOI issues, added documentation and contribution guidelines in main README.md.

When changes are made to the matlab-instrument-class, continuous integration tests are performed on the abstract classes, which are further inherited by the other instruments classes but cannot be applied to any instance class unless the corresponding instrument is actually hooked up.

There is no automatically generated code (e.g., .mlapp) included, such as code from MATLAB App Designer or other GUI creation tools. All of the MATLAB code, including any functions related to user interfaces or graphical components, has been manually written, ensuring that it reflects original contributions and not auto-generated templates or GUI elements. The 17,329 lines of MATLAB code in the report are entirely the result of manual development.

With these changes we created a new release: matlab-instrument-control v0.1.1.
