### 2024-09-24 #7275 @editorialbot
[REVIEW]: A MATLAB-based Instrument Control (MIC) package for fluorescence imaging
**Submitting author:** @sajjad88 (<a href="http://orcid.org/0000-0002-6910-5199">Sajjad Khan</a>)
**Repository:** https://github.com/LidkeLab/matlab-instrument-control
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
