### 2024-09-25 LidkeLab/matlab-instrument-control#124 @raacampbell
Code organisation
This comment is associated with a [manuscript review](https://github.com/openjournals/joss-reviews/issues/7275).

This software project has a large number of related classes and organising them better would help with feature discovery and long term maintainability.  Specifically:

* [x] Placing all MATLAB code into a single folder called `code`, `src`, or whatever is deemed suitable. This way it is clearer what folder(s) users should be adding to the MATLAB path.
* [x] Refactor to organise classes in a [tree of namespace folders](https://uk.mathworks.com/help/matlab/matlab_oop/namespaces.html).

I realise the second point will involve a lot or refactoring of both this project and projects that depend upon it, however this is really is the gold standard way of structuring projects like this in MATLAB. It will pay for itself down the line and modern editors or IDEs make this sort of thing pretty painless. A structure along these sort of lines would make sense:

```
code/
code/+mic/
code/+mic/+camera
code/+mic/+camera/@abstract
code/+mic/+camera/@example
code/+mic/+camera/@DCAM4Camera
code/+mic/+linearstage
code/+mic/+linearstage/@abstract
...
```

Following this, you call at the command line as follows:
```
mDCAM = mic.camera.DCAM4Camera;
```

This will help with feature discovery, since classes such as [MIC_ActiveReg3D](https://github.com/LidkeLab/matlab-instrument-control/blob/main/%40MIC_ActiveReg3D/MIC_ActiveReg3D.m), which is a much higher level entity, will not be lumped in with all the lower level stuff and abstract classes.
