### 2024-09-25 LidkeLab/matlab-instrument-control#126 @raacampbell
Provide a list of similar or related projects
This comment is associated with a https://github.com/openjournals/joss-reviews/issues/7275.

It would be great to have a list of similar or related MATLAB projects in the README. [lsmacq](https://github.com/danionella/lsmaq) comes to mind, but there must be others.
