### 2024-09-25 LidkeLab/matlab-instrument-control#127 @raacampbell
Provide a list of projects using the MIC
This comment is associated with a https://github.com/openjournals/joss-reviews/issues/7275.

Please include in the README a list of projects that uses the MIC.
