### 2024-09-25 LidkeLab/matlab-instrument-control#128 @raacampbell
Remove .DS_Store files
This comment is associated with a https://github.com/openjournals/joss-reviews/issues/7275.

There are numerous `.DS_Store` files from Macs. I would suggest to remove and stop them ever coming back with the git ignore file.
