### 2024-09-25 LidkeLab/matlab-instrument-control#129 @raacampbell
Project requires more extensive documentation, including examples
This comment is associated with a https://github.com/openjournals/joss-reviews/issues/7275.

There is an impressive array of classes here, which can be used for all sorts of different purposes. Knowing where to start will be tricky for new users and there is minimal documentation for those starting out. One of the key determinants for whether a project takes off and is used by others is the quality of the documentation and range of available examples. It's really worth taking time for this otherwise potential new users will just reinvent the wheel. I make the following suggestions:

* [ ] Move documentation out to a more suitable external site. e.g. create a GitHub Wiki for this repo or use any other documentation site of your choice,
* [ ] See #127
* [ ] See #126
* [ ] Include a few worked examples showing how to achieve sample tasks, such as setting up a linear stage and a camera to make a simple tile-scanning program.
* [ ] Document how the code is organised with diagrams so new contributors can more quickly understand how to make new concrete classes from the abstract classes. e.g. What does  `MIC_Abstract` contribute to `MIC_Camera_Abstract`.
* [ ] You can merge into this documentation site stuff that is currently scattered around the repository. e.g. the contribution instructions (which are great to see!)

## Example of related projects
For reference: [pyLabLib](https://pylablib.readthedocs.io/en/latest/index.html) is a similar project with a lot more documentation.
