### 2024-09-25 LidkeLab/matlab-instrument-control#130 @raacampbell
"Example" files are a fantastic idea but the naming is slightly confusing 
This comment is associated with a https://github.com/openjournals/joss-reviews/issues/7275.

It is an excellent idea to have the "Example" classes, as these allow users to construct programs using simulated hardware. They also can be used for unit testing downstream applications. The naming is a little confusing, though. I realise that probably the naming is related to the fact that these files are intended to be examples showing how to make a concrete class from an abstract class. However, all the classes you present here that inherit from an abstract class constitute a good example of how to perform this (assuming they are well documented).

I would therefore suggest renaming the "example" to "simulated" and integrate these into the example usage described in Issue #129. This would create more robust documentation and make it clearer what these classes can really achieve.
