### 2024-09-25 LidkeLab/matlab-instrument-control#131 @raacampbell
clc in a method is rather aggressive
This comment is associated with a https://github.com/openjournals/joss-reviews/issues/7275.

I see in [MIC_AndorCameraZyla.m](https://github.com/LidkeLab/matlab-instrument-control/blob/main/%40MIC_AndorCameraZyla/MIC_AndorCameraZyla.m) line 277 a `clc` command. These may well also be elsewhere, but I have not searched yet.

`clc` is somewhat aggressive. Your end-user ,might not want the command line cleared each time they start scanning (the command comes at the top of `start_scan`.

These commands ought to be removed so the end user who integrates MIC into their code can decide when/if this needs doing.
