### 2024-09-26 920e721
Fix bug that was causing autoName to produce invalid variable names
