### 2024-09-26 LidkeLab/matlab-instrument-control#133 @raacampbell
Tidy properties in concrete classes that inherit from an abstract class
This comment is associated with a https://github.com/openjournals/joss-reviews/issues/7275.

I look at `MIC_Camera_Abstract` and it declares a bunch of properties. Most or all of these are then declared again in the concrete classes, such as `MIC_DCAM4Camera`. It is confusing if stuff is apparently declared twice. If this is being done to avoid having to define abstract properties to get a concrete class, then why not just do not set the optional properties as Abstract? This is preferable to the apparent redundancy there currently is.

Likely the same thing is happening in other classes, but I didn't check.
