### 2024-09-26 LidkeLab/matlab-instrument-control#134 @raacampbell
More information needed in documentation for MIC_GalvoAnalog
This comment is associated with a https://github.com/openjournals/joss-reviews/issues/7275.

I would have expected a class that controls a galvo is involved somehow in making waveforms. I was confused, therefore, to discover the class just sets the galvo to a single fixed position. It would be helpful if this was explicit in the readme file and perhaps indicate what this is used for, since this class can not be used for scanning a beam over a sample.
