### 2024-09-26 a280045
Tidy indenting and spacing in gui
