+++ @MIC_PowerMeter_Abstract/gui.m
@@ -1,165 +1,191 @@
 function guiFig=gui(obj)
-% gui selects a data set from the pop-up menu, then
-% click one of the plot-type push buttons. Clicking the button
-% plots the selected data in the axes.
+    % gui selects a data set from the pop-up menu, then
+    % click one of the plot-type push buttons. Clicking the button
+    % plots the selected data in the axes.
 
-%prevent opening more that one gui for an objext.
-if ishandle(obj.GuiFigure)
-    guiFig = obj.GuiFigure;
-    figure(obj.GuiFigure);
-    return
-end
-        %  Create and then hide the GUI as it is being constructed.
-         guiFig = figure('Visible','on','Position',[250,120,1010,790]);
-        % Construct the components.
-        % editable text
-         hedit0 = uicontrol('Style','edit','Position',[175,730,200,45],'FontSize',20,'FontWeight','bold');
-         % static text
-         htext0 = uicontrol('Style','text','String','Current',...
-                   'Position',[100,720,70,45],'FontSize',12);
-         %static text
-         htext00 = uicontrol('Style','text','String','mW',...
-                   'Position',[380,720,30,45],'FontSize',12);      
-         %editable text      
-         hedit01 = uicontrol('Style','edit','Position',[510,730,200,45],'FontSize',20,'FontWeight','bold');
-         %static text
-         htext01 = uicontrol('Style','text','String','Max',...
-                   'Position',[450,720,50,45],'FontSize',12); 
-         %editable text      
-         htext011 = uicontrol('Style','text','String','mW',...
-                   'Position',[715,720,30,45],'FontSize',12); 
-        % statics text   
-         Htext1String=sprintf('%d<lambda(nm)<%d',obj.Limits(1),obj.Limits(2));
-        % static text
-         htext1 = uicontrol('Style','text','String',Htext1String,...
-                   'Position',[1100,680,200,30],'FontSize',12);
-        %editable text
-         hedit1 = uicontrol('Style','edit','Position',[800,649,150,30],'FontSize',12);
-        %static text 
-         htext2  = uicontrol('Style','text','String','nm',...
-                   'Position',[955,654,25,20],'FontSize',12);       
-        % push botton
-        hStartplot  = uicontrol('Style','pushbutton',...
-                    'String','Start Plot','Position',[800,570,80,30],...
-                    'FontSize',12,'Callback',@startPlotbutton_Callback);
-        % push button        
-        hStopplot  = uicontrol('Style','pushbutton',...
-                    'String','Stop Plot','Position',[890,570,80,30],...
-                    'FontSize',12,'Callback',@stopPlotbutton_Callback);
-        % editable text        
-        hedit3 = uicontrol('Style','edit','Position',[890,530,80,30],'FontSize',12);
-        % static text
-        htext3  = uicontrol('Style','text','String','T(S)',...
-                   'Position',[815,530,50,25],'FontSize',12); 
-        % push button        
-        hget  = uicontrol('Style','pushbutton',...
-                    'String','Get','Position',[473,460,175,30], ...
-                    'FontSize',12,'Callback',{@getbutton_Callback});
-        % editable       
-        hedit2 = uicontrol('Style','edit','Position',[800,420,140,30],'FontSize',12);
-        % static text
-        htext4 = uicontrol('Style','text','String','mW','Position',[940,415,40,30],'FontSize',12);
-        % popup menu 
-        hpopup = uicontrol('Style','popupmenu',...
-                  'String',{'Power(mW)','Temp(C)'},...
-                  'Position',[815,350,140,30],...
[…]
