### 2024-09-26 f71e2dc
Redundant code: GUI is visible by default anyway.
