+++ @MIC_PM100D/MIC_PM100D.m
@@ -1,4 +1,4 @@
-classdef MIC_PM100D < MIC_PowerMeter_Abstract 
+classdef MIC_PM100D < MIC_PowerMeter_Abstract
     % MIC_PM100D: Matlab Instrument class to control power meter PM100D.
     %
     % ## Description
@@ -7,13 +7,13 @@ classdef MIC_PM100D < MIC_PowerMeter_Abstract
     % set for power measurement, where the range of the wavelength is
     % 400nm < Lambda < 1100nm. The gui shows a movie of the plot of the
     % measured power where the shown period can be modified. It also shows
-    % the current power and the maximum measured power. To run this code 
+    % the current power and the maximum measured power. To run this code
     % you need the power meter to be connected to the machine.
     %
     % ## Constructor
     % Example: P = MIC_PM100D; P.gui
     %
-    % ## Key Functions: 
+    % ## Key Functions:
     % constructor(), exportState(), send(), minMaxWavelength(), getWavelength(), measure(), setWavelength(), shutdown()
     %
     % ## REQUIREMENTS:
@@ -23,24 +23,26 @@ classdef MIC_PM100D < MIC_PowerMeter_Abstract
     %    MIC_PowerMeter_Abstract.m
     %
     % ### CITATION: Mohamadreza Fazel, Lidkelab, 2017.
-    
+
     properties (SetAccess=protected)
         InstrumentName='PM100D'; %Name of the instrument.
     end
+
+
     methods (Static)
         function obj=MIC_PM100D
             %This is the constructor.
             %example PM = MIC_PM100D
-            obj@MIC_PowerMeter_Abstract(~nargout);    
-           
+            obj@MIC_PowerMeter_Abstract(~nargout);
+
             % Find a VISA-USB object.
             vendorinfo = instrhwinfo('visa','ni');
             s=vendorinfo.ObjectConstructorName{1};
-            
+
             if isempty(obj.VisaObj)
                 obj.VisaObj = eval(s);
             else
-                
+
                 fclose(obj.VisaObj);
                 clear obj.VisaObj;
                 obj.VisaObj = eval(s);
@@ -49,59 +51,62 @@ classdef MIC_PM100D < MIC_PowerMeter_Abstract
             fopen(obj.VisaObj);
             % Measure the limits of the wavelength.
             obj.Limits=minMaxWavelength(obj);
-        end 
-      function unitTest()
-          %testing the class.
-          try
-          TestObj=MIC_PM100D();
-          fprintf('Constructor is run and an object of the class is made.\n');
-          Limit = minMaxWavelength(TestObj);
-          fprintf('Min wavelength: %d, Max wavelength: %d\n', Limit(1),Limit(2));
-          getWavelength(TestObj);
-          fprintf('The current wavelength is %d nm.\n',TestObj.Lambda);
-          TestObj.Lambda=600;
-          setWavelength(TestObj);
-          fprintf('The wavelength is set to 600 nm.\n');
-          State=exportState(TestObj);
-          fprintf('The exportState function was successfully tested.\n');
-          TestObj.delete();
-          fprintf('The port is closed and the object is deleted.\n');
-          fprintf('The class is successfully tested :)\n')
-          catch E
-             fprintf('Sorry an error has ocured :(\n'); 
-              E
-          end
-      end
-    end
-    
+
+            obj.Ask = 'power'; % by default we read power
+        end %constructor
+
+        function unitTest()
+            %testing the class.
+            try
+                TestObj=MIC_PM100D();
+                fprintf('Constructor is run and an object of the class is made.\n');
+                Limit = minMaxWavelength(TestObj);
[…]
