### 2024-09-26 2862187
fail gracefully if power meter is not plugged in
