### 2024-09-26 bf2ea2b
do not need "clear obj" this happens anyway
