### 2024-09-26 b5a2f79
Made it possible to run setWavelength method with an input argument
