### 2024-09-26 815e620
Add measurePower() and measureTemperature() methods because running measure() at the CLI is otherwise not convenient
