### 2024-09-26 e34eb9c
make object calls more canonical and replace with new power measurement method
