### 2024-09-27 466764b
Remove odd if statement from constructor. Not possible to ever end up here.
