### 2024-09-27 f0192d2
Use the usual notation for calling object methods
