### 2024-09-27 81ea49b
Honor sensor limits not hard-coded limits
