### 2024-09-27 4adddb6
Use more universal wavelength setting commands. Code now verified to work with PM101 as well!
