### 2024-09-27 8580535
read a few more device settings from power meter
