### 2024-09-27 da67e48
add autorange methods
