### 2024-09-27 0876887
minor bugfix: do not run close if nothing was connected
