+++ src/+mic/@CavroSyringePump/CavroSyringePump.m
@@ -1,0 +1,179 @@
+classdef MIC_CavroSyringePump < MIC_Abstract
+% MIC_CavroSyringePump Class 
+% 
+% ## Description
+% The `MIC_CavroSyringePump` class controls a Cavro syringe pump via USB, specifically designed for 
+% pump PN 20740556 -D. This class may work with other Cavro brand syringe pumps but has only been tested with the 
+% specified model. It can perform any operation described in the Cavro XP3000 operators manual (e.g. in Appendix G - Command Quick Reference). 
+% 
+% ## Installation Requirements
+% - MATLAB R2014b or later (R2017a or later recommended)
+% - Operating System: Windows (modifications required for UNIX systems, particularly in serial port behaviors)
+% - Dependency: `MIC_Abstract.m`
+% 
+% ##  Functions: 
+% delete, exportState, updateGui, gui, connectSyringePump,
+% readAnswerBlock, executeCommand, reportCommand,
+% querySyringePump, cleanAnswerBlock, decodeStatusByte, unitTest
+% 
+% ## Usage
+% ```matlab
+% % Create an instance of the Cavro syringe pump controller
+% Pump = MIC_CavroSyringePump();
+% % Connect to the pump
+% [Message, Status] = Pump.connectSyringePump();
+% % Execute a command to move the plunger
+% Pump.executeCommand('Move Plunger to 1000');
+% % Check the pump's status
+% Pump.querySyringePump();
+% % Disconnect and cleanup
+% delete(Pump);
+% ```   
+% ### CITATION: David Schodt, Lidke Lab, 2018
+
+    properties
+        DeviceAddress = 1; % ASCII address for device
+        DeviceSearchTimeout = 10; % timeout(s) to search for a pump
+        DeviceResponseTimeout = 10; % timeout(s) for valid device response
+        SerialPort = 'COM3'; % default COM port
+    end
+    
+    
+    properties (SetAccess = protected) % users shouldn't set these
+        InstrumentName = 'CavroSyringePump'; % name of the instrument
+        SyringePump; % serial object for the connected syringe pump
+        PlungerPosition; % absolute plunger position (0-3000)
+        ReadableAction; % activity of the pump/pump response to a report
+        VelocitySlope; % ramping slope of plunger from StartVelocity->Top
+        StartVelocity; % plunger starting velocity (half-steps/sec)
+        TopVelocity; % max plunger velocity (half-steps/sec)
+        CutoffVelocity; % plunger stop velocity (half-steps/sec)
+    end
+    
+    
+    properties (SetObservable)
+        StatusByte = 0; % status of the pump, 0 if not connected
+    end
+    
+    
+    properties (Hidden)
+        MatlabRelease; % version of MATLAB that is using this class
+        StartGUI = false;
+    end
+    
+    
+    properties (Dependent) % determined on demand
+        ReadableStatus; % user readable status of the syringe pump
+    end
+       
+    
+    methods
+        function obj = MIC_CavroSyringePump()
+            %Constructor for the Cavro syringe pump object.
+                       
+            % If needed, automatically assign a name to the instance of
+            % this class (i.e. if user forgets to do this).
+            obj = obj@MIC_Abstract(~nargout);
+            
+            % Set property listener(s). 
+            addlistener(obj, 'StatusByte', 'PostSet', ...
+                @obj.statusByteChange);
+            
+            % Determine the current version of MATLAB in use.
+            obj.MatlabRelease = version('-release');
+        end
+        
+        function delete(obj)
+            %Defines a class destructor for the syringe pump.
+            
+            % If the serial object obj.SyringePump exists (i.e. a syringe
+            % pump had been connected to), close the connection to it and
+            % clean up.
+            if ~isempty(obj.SyringePump)
+                fclose(obj.SyringePump); % close serial connection
[…]
