+++ src/+mic/+camera/@IMGSourceCamera/IMGSourceCamera.m
@@ -1,0 +1,440 @@
+classdef MIC_IMGSourceCamera <  MIC_Camera_Abstract
+    % MIC_IMGSourceCamera: Matlab instument class for ImagingSource camera.
+    %
+    % ## Description 
+    % It requires dll to be registered in MATLAB.
+    % TISImaq interfaces directly with the IMAQ Toolbox. This allows you to 
+    % bring image data directly into MATLAB for analysis, visualization, 
+    % and modelling.
+    % The plugin allows the access to all camera Properties as they are 
+    % known from IC Capture. The plugin works in all Matlab versions 
+    % since 2013a. Last tested version is R2016b.
+    % After installing the plugin it must be registered in Matlab manually.
+    % How to do that is shown in a Readme, that can be displayed after 
+    % installation. 
+    % imaqregister('C:\Program Files (x86)\TIS IMAQ for MATLAB R2013b\x64\TISImaq_R2013.dll')
+    % http://www.theimagingsource.com/support/downloads-for-windows/extensions/icmatlabr2013b/
+    % This was done with imaqtool using Tools menu.
+    %
+    % ## Contructor
+    % Example: obj=MIC_IMGSourceCamera();
+    %
+    % ## Key Functions:
+    % delete, shutdown, exportState
+    %
+    % ## REQUIREMENTS: 
+    %   MIC_Abstract.m
+    %   MIC_Camera_Abstract.m
+    %   MATLAB software version R2013a or later
+    %
+    % ### CITATION: , Lidkelab, 2017.
+    
+    properties(Access=protected)
+        AbortNow;           %stop acquisition flag
+        FigurePos;
+        FigureHandle;
+        ImageHandle;
+        ReadyForAcq=0;        %If not, call setup_acquisition
+        TextHandle;
+        CameraHandle;
+        CameraCap;
+    end
+    
+    properties (Hidden)
+        StartGUI=false;     %Defines GUI start mode.  'true' starts GUI on object creation.
+    end
+    
+    properties(SetAccess=protected)
+        CameraIndex;        %index used when more than one camera
+        ImageSize;          %size of current ROI
+        LastError;          %last errorcode
+        Manufacturer;       %camera manufacturer
+        Model;              %camera model
+        CameraParameters;   %camera specific parameters
+        CameraSetting;
+        XPixels;            %number of pixels in first dimention
+        YPixels;            %number of pixels in second dimention
+        InstrumentName='IMGSourceCamera'
+    end
+    
+    properties
+        Binning;            %   [binX binY]        
+        Data;               %   last acquired data
+        ExpTime_Focus;      %   focus mode exposure time
+        ExpTime_Capture;    %   capture mode exposure time
+        ExpTime_Sequence;   %   sequence mode expsoure time
+        ROI;                %   [Xstart Xend Ystart Yend]
+        SequenceLength=1;   %   Kinetic Series length
+        SequenceCycleTime;  %   Kinetic Series cycle time (1/frame rate)
+        ScanMode;           %   scan mode for Hamamatsu sCMOS camera
+        TriggerMode;        %   trigger mode for Hamamatsu sCMOS camera
+        DefectCorrection;   %   defect correction  for Hamamatsu sCMOS camera
+        GuiDialog;
+    end
+    
+    methods
+        
+        function obj=MIC_IMGSourceCamera() 
+            % Object constructor
+            obj = obj@MIC_Camera_Abstract(~nargout);
+        end
+        
+        function abort(obj)
+            % Abort function
+            stop(obj.CameraHandle);
+            obj.ReadyForAcq=0;
+        end
+        
+        function shutdown(obj)
+            % Object shutdown
+            delete(obj.CameraHandle);
+            clear obj.CameraHandle;
+        end
+        
[…]

[…]
