### 2024-10-02 3232b9e
enamed abstract; moved some cameras to +camera; more reorg naming
