+++ src/+mic/@DynamixelServo/DynamixelServo.m
@@ -1,5 +1,5 @@
-classdef MIC_DynamixelServo < MIC_Abstract
-    % MIC_DynamixelServo: Matlab Instrument Class for Dynamixel Servos
+classdef DynamixelServo < mic.abstract
+    % mic.DynamixelServo: Matlab Instrument Class for Dynamixel Servos
     % 
     % ## Description
     %   Dynamixel Servos are used to control the rotation of filter wheels
@@ -8,7 +8,7 @@ classdef MIC_DynamixelServo < MIC_Abstract
     %   Setup.doc
     %
     %   ## Usage Example:
-    %           obj=MIC_DynamixelServo(ServoId,Port,Bps);
+    %           obj=mic.DynamixelServo(ServoId,Port,Bps);
     %           ServoId: Id of servo(is written on servo)
     %           Port: COM port to which servo is connected (Optional)
     %           Bps: Baud setting for port (Optional)
@@ -23,7 +23,7 @@ classdef MIC_DynamixelServo < MIC_Abstract
     %
     %   ## REQUIRES:
     %     Matlab 2014b or higher
-    %     MIC_Abstract.m
+    %     mic.abstract.m
     %     Roboplus software
     %     Driver library for servo
     %     Driver library for USB2Dynamixel
@@ -128,18 +128,18 @@ classdef MIC_DynamixelServo < MIC_Abstract
     end
     
     methods
-        function obj=MIC_DynamixelServo(Id,Port,Bps)
+        function obj=DynamixelServo(Id,Port,Bps)
             % Object constructor
-            obj = obj@MIC_Abstract(~nargout);
+            obj = obj@mic.abstract(~nargout);
             if ~exist('Id','var') || Id < 1 || Id > 255
-                error('MIC_DynamixelServo:Id','Invalid servo Id');
+                error('mic.DynamixelServo:Id','Invalid servo Id');
             end
             obj.Id = Id;          
             if ~exist('Bps','var')
-                Bps = MIC_DynamixelServo.DEFAULT_BAUDNUM;
+                Bps = DynamixelServo.DEFAULT_BAUDNUM;
             end
             obj.Bps = Bps;            
-            MIC_DynamixelServo.loadlibrary();
+            mic.DynamixelServo.loadlibrary();
             
             %Try to find correct com port if not inputed
             if exist('Port','var')
@@ -148,7 +148,7 @@ classdef MIC_DynamixelServo < MIC_Abstract
                obj.Port=obj.findport(Id,Bps);
             end
             if obj.ping() ~= obj.COMM_RXSUCCESS
-                error('MIC_DynamixelServo:MissingServo','Could not connect to servo at Id:%d Port:%d Bps:%d',Id,obj.Port,Bps);
+                error('mic.DynamixelServo:MissingServo','Could not connect to servo at Id:%d Port:%d Bps:%d',Id,obj.Port,Bps);
             end
             
             % set Led and MovingSpeed to allowed values
@@ -157,7 +157,7 @@ classdef MIC_DynamixelServo < MIC_Abstract
         end
         
         function delete(obj)
-            %MIC_DynamixelServo destructor
+            %mic.DynamixelServo destructor
             obj.shutdown();
         end
         
@@ -169,10 +169,10 @@ classdef MIC_DynamixelServo < MIC_Abstract
         function checkCommStatus(obj)
             % Checks status of COM port
             CommStatus = int32(obj.callDynamixel('dxl_get_result'));
-            if CommStatus == MIC_DynamixelServo.COMM_RXSUCCESS
-                MIC_DynamixelServo.printErrorCode();
+            if CommStatus == DynamixelServo.COMM_RXSUCCESS
+                mic.DynamixelServo.printErrorCode();
[…]
