+++ src/+mic/@GalvoDigital/gui.m
@@ -1,10 +1,10 @@
-% gui: Graphical User Interface to MIC_GalvoDigital
+% gui: Graphical User Interface to mic,GalvoDigital
 %
 % Functions: gui2properties, properties2gui, closeFigure, CalAngle,
 %            setAngle_Button, setParameters_Button, ToggleGalvo
 %
 % REQUIREMENTS:
-%    MIC_Abstract.m
+%    mic.abstract.m
 %    MATLAB NI-DAQmx driver installed via the Support Package Installer
 %
 % CITATION: Hanieh Mazloom-Farsibaf, Lidlelab, 2017.
