+++ src/+mic/@OptotuneLens/gui.m
@@ -1,5 +1,5 @@
 function guiFig = gui(obj)
-%gui Graphical User Interface to MIC_OptotuneLens
+%gui Graphical User Interface to mic.OptotuneLens
 
 %Prevent opening more than one figure for same instrument
 if ishandle(obj.GuiFigure)

[…]
