+++ src/SAVE/+3Dstage/@example/Example_3DStage.m
@@ -1,381 +1,0 @@
-classdef Example_3DStage< MIC_3DStage_Abstract
-    % This class is an example implementation of 3D Stage Class.
-    % This class simulates a 3D stage that can move along x,y,z axes.
-    
-    % REQUIRES:
-    % MIC_3DStage_Abstract.m
-    
-    % Documentation: For detailed documentation check Readme.md file.
-    
-    % CITATION: Sajjad Khan, Lidkelab, 2024.
-    
-    properties (SetAccess = protected)
-        InstrumentName = 'Simulated 3D Stage'; % Name of the instrument
-        Position = [0, 0, 0];   % Example position [x, y, z]
-        PositionUnit = 'mm';    % Example position unit (millimeters)
-    end
-    
-    properties (Hidden)
-        StartGUI = false;                         % GUI does not start automatically
-    end
-    
-    methods
-        function obj = Example_3DStage()
-            % Constructor
-            obj@MIC_3DStage_Abstract(~nargout);
-        end
-        
-        function center(obj)
-            % Method to set the stage to the center (0, 0, 0)
-            obj.Position = [0, 0, 0];
-            fprintf('Stage positioned at the center.\n');
-        end
-        
-        function setPosition(obj, position)
-            % Method to set the stage to a specified position
-            if numel(position) == 3
-                obj.Position = position;
-                fprintf('Stage positioned at: [%f, %f, %f] %s\n', position, obj.PositionUnit);
-            else
-                error('Position must be a 3-element vector [x, y, z]');
-            end
-        end
-        
-        function [Attributes, Data, Children] = exportState(obj)
-            % Method to export the state of the stage
-            Attributes = struct('PositionUnit', obj.PositionUnit);
-            Data = struct('Position', obj.Position);
-            Children = struct(); % No children in this example
-        end
-        
-        function gui(obj)
-            %gui Graphical User Interface to MIC_3DStage_Abstract
-            
-            h = findall(0,'tag','MIC_3DStage_Abstract_gui');
-            %Prevent opening more than one figure for same instrument
-            if ~(isempty(h))
-                figure(h);
-                return;
-            end
-            
-            
-            %Set the position of figure in windows
-            xsz=300;
-            ysz=400;
-            xst=100;
-            yst=100;
-            bszx=75;
-            bszy=30;
-            txsz = 75;
-            txszy = 20;
-            etxsz = 50;
-            PositionUnit=obj.PositionUnit;
-            
-            %Open figure
-            guiFig = figure('Units','pixels','Position',[xst yst xsz ysz],...
-                'MenuBar','none','ToolBar','none','Visible','on',...
-                'NumberTitle','off','UserData',0,...
-                'Tag','MIC_3DStage_Abstract_gui','HandleVisibility','off');
-            defaultBackground = get(0,'defaultUicontrolBackgroundColor');
-            set(guiFig,'Color',defaultBackground)
-            
-            %Initialize GUI properties
-            guiFig.WindowButtonDownFcn=@properties2gui;
-            
-            
-            %top
-            uicontrol('Parent',guiFig, 'Style', 'text', 'String', ...
-                ['X Position (' PositionUnit ')'],'Position', [10 ysz-(bszy+10) txsz bszy]);
-            uicontrol('Parent',guiFig, 'Style', 'text', 'String', ...
-                ['Y Position (' PositionUnit ')'],'Position', [10 ysz-2*(bszy+10) txsz bszy]);
-            uicontrol('Parent',guiFig, 'Style', 'text', 'String', ...
-                ['Z Position (' PositionUnit ')'],'Position', [10 ysz-3*(bszy+10) txsz bszy]);
-            
[…]
