### 2024-10-02 3de05c4
Finished the linearstage routines.
