+++ src/+mic/+linearstage/@MCLMicroDrive/moveDistance.m
@@ -19,13 +19,13 @@ end
 % Validate the input parameters.
 if ((obj.Velocity<obj.VelocityBounds(1)) ...
         || (obj.Velocity>obj.VelocityBounds(2)))
-    error(['MIC_MCLMicroDrive.moveDistance(): invalid velocity %g ', ...
+    error(['mic.linearstage.MCLMicroDrive.moveDistance(): invalid velocity %g ', ...
         'um/s - velocity must be within %g um/s and %g um/s'], ...
         obj.Velocity, obj.VelocityBounds(1), obj.VelocityBounds(2));
 end
 Distance = double(Distance);
 if (isnan(Distance) || isinf(Distance))
-    error('MIC_MCLMicroDrive.moveDistance(): invalid distance %g', ...
+    error('mic.linearstage.MCLMicroDrive.moveDistance(): invalid distance %g', ...
         Distance)
 end
 

[…]
