### 2024-10-03 aaf53ec
Finsiged converting mic.lightsource classes.
