### 2024-10-03 6b86f1e
1st pass finishing camera namespace.
