+++ src/+mic/+stage3D/@NanoMaxPiezos/NanoMaxPiezos.m
@@ -1,8 +1,8 @@
-classdef MIC_NanoMaxPiezos < MIC_3DStage_Abstract
-    % MIC_NanoMaxPiezos Class
+classdef NanoMaxPiezos < mic.stage3D.abstract
+    % mic.stage3D.NanoMaxPiezos Class
     %
     % ## Description
-    % The `MIC_NanoMaxPiezos` class provides comprehensive control over the three piezo stages (x, y, z) of a Thorlabs NanoMax stage. It is designed to handle precise positioning necessary in advanced microscopy setups.
+    % The `mic.stage3D.NanoMaxPiezos` class provides comprehensive control over the three piezo stages (x, y, z) of a Thorlabs NanoMax stage. It is designed to handle precise positioning necessary in advanced microscopy setups.
     %
     % ## Key Features
     % - Individual control of x, y, and z piezo stages.
@@ -11,16 +11,16 @@ classdef MIC_NanoMaxPiezos < MIC_3DStage_Abstract
     % - Integration with both TCube and KCube piezo systems.
     %
     % ## Requirements
-    % - `MIC_Abstract.m`
-    % - `MIC_3DStage_Abstract.m`
-    % - Piezo control classes (`MIC_TCubePiezo.m` and `MIC_KCubePiezo.m`)
+    % - `mic.abstract.m`
+    % - `mic.stage3D.abstract.m`
+    % - Piezo control classes (`mic.linearstage.TCubePiezo.m` and `mic.linearstage.KCubePiezo.m`)
     % - MATLAB 2016b or later.
     %
     % ## Installation Notes
     % Before using this class, ensure that all dependent classes and required Thorlabs drivers are installed and properly configured on your system.
     %
     % ## Methods
-    % ### `Constructor (MIC_NanoMaxPiezos())`
+    % ### `Constructor (mic.stage3D.NanoMaxPiezos())`
     % Initializes piezo controllers for x, y, and z axes based on provided serial numbers. It attempts to connect to the piezos, with error handling to manage connection issues.
     %
     % ### `center()`
@@ -38,7 +38,7 @@ classdef MIC_NanoMaxPiezos < MIC_3DStage_Abstract
     % ## Usage Example
     % ```matlab
     % % Initialize NanoMax Piezos
-    % nmPiezos = MIC_NanoMaxPiezos('81850186', '84850145', '81850193', '84850146', '81850176', '84850203', 3);
+    % nmPiezos = mic.stage3D.NanoMaxPiezos('81850186', '84850145', '81850193', '84850146', '81850176', '84850203', 3);
     %
     % % Center all piezos
     % nmPiezos.center();
@@ -79,7 +79,7 @@ classdef MIC_NanoMaxPiezos < MIC_3DStage_Abstract
     end
     
     methods
-        function obj = MIC_NanoMaxPiezos(...
+        function obj = NanoMaxPiezos(...
                 ControllerXSerialNum, StrainGaugeXSerialNum, ...
                 ControllerYSerialNum, StrainGaugeYSerialNum, ...
                 ControllerZSerialNum, StrainGaugeZSerialNum, ...
@@ -88,7 +88,7 @@ classdef MIC_NanoMaxPiezos < MIC_3DStage_Abstract
             
             % If needed, automatically assign a name to the instance of
             % this class (i.e. if user forgets to do this).
-            obj = obj@MIC_3DStage_Abstract(~nargout);
+            obj = obj@mic.stage3D.abstract(~nargout);
             
             % Set the object properties based on the appropriate inputs.
             if exist('MaxPiezoConnectAttempts', 'var')
@@ -137,12 +137,12 @@ classdef MIC_NanoMaxPiezos < MIC_3DStage_Abstract
[…]
