+++ src/+mic/@Example_LightSource/Example_LightSource.m
@@ -1,0 +1,217 @@
+classdef Example_LightSource < mic.lightsource.abstract
+    % This class is an example implementation of mic.lightsource.abstract.
+    % This class provides full functionalities for a simulated light source such as Laser.
+    
+    % REQUIRES:
+    % mic.lightsource.abstract.m
+    
+    % Documentation: For detailed documentation check Readme.md file.
+    
+    % CITATION: Sajjad Khan, Lidkelab, 2024.
+    
+    properties (SetAccess = protected)
+        InstrumentName = 'ExampleLightSource'; % Name of the instrument
+        PowerUnit = 'Watts';
+        Power = 0;  % Initialize Power to a valid scalar within the range
+        IsOn = 0;
+        MinPower = 0;
+        MaxPower = 100;
+    end
+    
+    
+    properties (Hidden)
+        StartGUI = false; % GUI does not start automatically by default
+    end
+    
+    methods
+        function obj = Example_LightSource()
+            obj = obj@mic.lightsource.abstract(~nargout);
+        end
+        
+        function setPower(obj, power)
+            if power < obj.MinPower || power > obj.MaxPower
+                error('Power value must be between %d and %d %s.', obj.MinPower, obj.MaxPower, obj.PowerUnit);
+            end
+            obj.Power = power;
+            fprintf('Power set to %g %s\n', obj.Power, obj.PowerUnit);
+        end
+        
+        function on(obj)
+            if obj.Power <= obj.MinPower
+                error('Power must be set above minimum power to turn on.');
+            end
+            obj.IsOn = 1;
+            fprintf('Light source turned on.\n');
+        end
+        
+        function off(obj)
+            obj.IsOn = 0;
+            fprintf('Light source turned off.\n');
+        end
+        
+        function shutdown(obj)
+            obj.off();
+            fprintf('Light source is shutting down.\n');
+        end
+        
+        function guiFig = gui(obj)
+            %gui Graphical User Interface to mic.lightsource.abstarct
+            
+            %Prevent opening more than one figure for same instrument
+            if ishandle(obj.GuiFigure)
+                guiFig = obj.GuiFigure;
+                figure(obj.GuiFigure);
+                return
+            end
+            
+            %Open figure
+            guiFig = figure('NumberTitle','off','Resize','off','Units','pixels','MenuBar','none',...
+                'ToolBar','none','Visible','on', 'Position',[100 100 450 300]);
+            
+            %Construct the components
+            handles.output = guiFig;
+            guidata(guiFig,handles);
+            
+            %set the basic prameters
+            minPower=obj.MinPower;
+            maxPower=obj.MaxPower;
+            unitPower=obj.PowerUnit;
+            
+            handles.sliderPower=uicontrol('Parent',guiFig,'Style','slider','Min',minPower,...
+                'Max',maxPower,'Value',minPower,'SliderStep',[0.1 0.1],...
+                'Position', [118 220 200 35],'Tag','positionSlider','Callback',@sliderfn);
+            handles.textMinPower = uicontrol('Style','text','String','Min Power',...
+                'Position',[35 240,80,15],'FontSize',10);
+            handles.valueMinPower = uicontrol('Style','text','String',[num2str(minPower),' ',unitPower],...
+                'Position',[48 220,70,20]);
+            handles.textMaxPower = uicontrol('Style','text','String','Max Power',...
+                'Position',[325 240,80,15],'FontSize',10);
+            handles.valueMaxPower = uicontrol('Style','text','String',[num2str(maxPower),' ',unitPower],...
+                'Position',[337 220,70,20]);
+            handles.textSetPower = uicontrol('Style','text','String','Set Power',...
[…]
