+++ src/+mic/@Example_Camera/Example_Camera.m
@@ -1,0 +1,272 @@
+classdef Example_Camera < mic.camera.abstract
+    % This is an example implementation of mic.camera.abstract 
+    % Matlab Instrument Control Camera Class. 
+    
+    % REQUIRES: 
+    % mic.abstract.m
+    
+    % Documentation: For detailed documentation check Readme.md file.
+    
+    % CITATION: Sajjad Khan, Lidkelab, 2024.
+
+    properties(SetAccess=protected)
+        InstrumentName = 'Simulated Camera';
+        CameraIndex = 1;
+        ImageSize = [1024, 768];
+        LastError = '';
+        Manufacturer = 'MyCam';
+        Model = 'CamModelX100';
+        CameraParameters = struct('Gain', 1, 'FrameRate', 30);
+        XPixels = 1024;
+        YPixels = 768;
+    end
+
+    properties
+        Binning = [1 1];
+        Data = [];
+        ExpTime_Focus = 0.01;
+        ExpTime_Capture = 0.02;
+        ExpTime_Sequence = 0.05;
+        ROI = [1 1024 1 768];
+        SequenceLength = 10;
+        SequenceCycleTime = 0.1;
+        TriggerMode = 'internal';
+    end
+
+    properties(Access=protected)
+        AbortNow = false;
+        FigurePos;
+        FigureHandle;
+        ImageHandle;
+        ReadyForAcq = false;
+        TextHandle;
+        TimerHandle; % Timer object handle
+    end
+    
+    properties (Hidden)
+        StartGUI = false; % GUI does not start automatically by default
+    end
+
+    methods
+        function obj = Example_Camera()
+            obj = obj@mic.camera.abstract(~nargout);
+        end
+
+        function state = exportState(obj)
+            state = obj.CameraParameters;
+            state.ImageSize = obj.ImageSize;
+            state.ExpTimes = [obj.ExpTime_Focus, obj.ExpTime_Capture, obj.ExpTime_Sequence];
+            state.ROI = obj.ROI;
+            state.Binning = obj.Binning;
+        end
+
+        function abort(obj)
+            obj.AbortNow = true;
+            disp('Abortion requested.');
+        end
+
+        function errorcheck(obj, funcname)
+            disp(['Checking errors for function: ', funcname]);
+            obj.LastError = '';
+        end
+
+        function out = getlastimage(obj)
+            disp('Retrieving last image.');
+            out = rand(obj.ImageSize); % Simulate an image
+        end
+
+        function out = getdata(obj)
+            disp('Getting data.');
+            out = rand(obj.ImageSize); % Simulate data acquisition
+            obj.Data = out;
+        end
+
+        function initialize(obj)
+            disp('Initializing camera settings.');
+            obj.ReadyForAcq = true;
+        end
+
+        function setup_acquisition(obj)
+            disp('Setting up acquisition parameters.');
+            if ~obj.ReadyForAcq
+                disp('Camera not ready. Initializing now.');
+                obj.initialize();
+            end
+        end
+        
+        function shutdown(obj)
+            disp('Shutting down the camera.');
+            obj.abort();
+            if ~isempty(obj.TimerHandle) && isobject(obj.TimerHandle) && isvalid(obj.TimerHandle)
+                stop(obj.TimerHandle);
+                delete(obj.TimerHandle);
+                obj.TimerHandle = [];
+            end
+            if ishandle(obj.FigureHandle)
+                close(obj.FigureHandle);
+                obj.FigureHandle = [];
+            end
+            disp('Camera shutdown completed.');
+        end
+        
+        function img = start_capture(obj)
+            disp('Starting capture mode.');
+            if obj.ReadyForAcq
+                disp('Camera is running in capture mode.');
[…]

[…]
