### 2024-10-07 5a05db8
Update MICclasses and add a hierarchical version.
