+++ doc/MICclassesHier.md
@@ -1,0 +1,64 @@
+### Matlab Instrument Control classes
+
+|**MIC** classes|description|
+-------------|---
+[mic.ActiveReg3D](../src/+mic/@ActiveReg3D/Readme.md) | active 3D registration using camera and stage control via piezoelectric actuators
+[mic.H5](../src/+mic/@H5/Readme.md) | static methods for working with HDF5 files
+[mic.Joystick](../src/+mic/@Joystick/Readme.md) | control the TIRF stage with a joystick
+[mic.abstract](../src/+mic/@abstract/Readme.md) | Matlab Instrument Control abstract class
+&nbsp;&nbsp;&nbsp;[mic.camera.abstract](../src/+mic/+camera/@abstract/Readme.md) | common interface for all cameras
+&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[mic.camera.AndorCamera](../src/+mic/+camera/@AndorCamera/Readme.md) | control an Andor camera
+&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[mic.camera.AndorCameraZyla](../src/+mic/+camera/@AndorCameraZyla/Readme.md) | control an Andor camera
+&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[mic.camera.DCAM4Camera](../src/+mic/+camera/@DCAM4Camera/Readme.md) | control a Hamamatsu camera using the DCAM4 API
+&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[mic.camera.HamamatsuCamera](../src/+mic/+camera/@HamamatsuCamera/Readme.md) | control a Hamamatsu camera
+&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[mic.camera.Imaq](../src/+mic/+camera/@Imaq/Readme.md) | class designed for camera control using the Image Acquisition Toolbox
+&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[mic.camera.IMGSourceCamera](../src/+mic/+camera/@IMGSourceCamera/Readme.md) | control an ImagingSource camera
+&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[mic.camera.PyDcam](../src/+mic/+camera/@PyDcam/Readme.md) | control a camera through a Python interface
+&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[mic.camera.ThorlabsIR](../src/+mic/+camera/@ThorlabsIR/Readme.md) | control a Thorlabs IR camera
+&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[mic.camera.ThorlabsSICamera](../src/+mic/+camera/@ThorlabsSICamera/Readme.md) | control a Thorlabs Scientific Camera via a USB port
+&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[mic.camera.IRSyringPump](../src/+mic/@IRSyringPump/Readme.md) | use the Syringe Pump at the same time data is taken with IRCamera
+&nbsp;&nbsp;&nbsp;[mic.lightsource.abstract](../src/+mic/+lightsource/@abstract/Readme.md) | abstract class for all light source devices
[…]
