### 2024-10-07 a3b9480
SOme small fixes for MICclassesHier.md
