+++ doc/MICclassesHier.md
@@ -16,7 +16,7 @@
 &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[mic.camera.PyDcam](../src/+mic/+camera/@PyDcam/Readme.md) | control a camera through a Python interface
 &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[mic.camera.ThorlabsIR](../src/+mic/+camera/@ThorlabsIR/Readme.md) | control a Thorlabs IR camera
 &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[mic.camera.ThorlabsSICamera](../src/+mic/+camera/@ThorlabsSICamera/Readme.md) | control a Thorlabs Scientific Camera via a USB port
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[mic.camera.IRSyringPump](../src/+mic/@IRSyringPump/Readme.md) | use the Syringe Pump at the same time data is taken with IRCamera
+&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[mic.camera.IRSyringPump](../src/+mic/+camera/@IRSyringPump/Readme.md) | use the Syringe Pump at the same time data is taken with IRCamera
 &nbsp;&nbsp;&nbsp;[mic.lightsource.abstract](../src/+mic/+lightsource/@abstract/Readme.md) | abstract class for all light source devices
 &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[mic.lightsource.CoherentLaser561](../src/+mic/+lightsource/@CoherentLaser561/Readme.md) | coherent Sapphire laser 561
 &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[mic.lightsource.CrystaLaser405](../src/+mic/+lightsource/@CrystaLaser405/Readme.md) | CrystaLaser 405 nm
@@ -36,9 +36,11 @@
 &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[mic.linearstage.KCubePiezo](../src/+mic/+linearstage/@KCubePiezo/Readme.md) | control a Thorlabs KCube Piezo stage
 &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[mic.linearstage.MCLMicroDrive](../src/+mic/+linearstage/@MCLMicroDrive/Readme.md) | control a Mad City Labs Micro Stage
 &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[mic.linearstage.TCubePiezo](../src/+mic/+linearstage/@TCubePiezo/Readme.md) | control a Thorlabs TCube Piezo stage
-&nbsp;&nbsp;&nbsp;[mic.powermeter.abstract](../src/+mic/+powermeter.@abstract/Readme.md) |  control a power meter (PM100D)
+&nbsp;&nbsp;&nbsp;[mic.powermeter.abstract](../src/+mic/+powermeter/@abstract/Readme.md) |  control a power meter (PM100D)
 &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[mic.powermeter.PM100D](../src/+mic/+powermeter/@PM100D/Readme.md) | control the power meter PM100D
-&nbsp;&nbsp;&nbsp;[mic.APTPiezo](../src/+mic/@APTPiezo/Readme.md) | control the Thorlabs APT piezo
[…]
