### 2024-10-08 5b0e927
fixes of typos and such; expand MICclasses to include Examples.
