+++ src/+mic/+stage3D/@abstract/gui.m
@@ -1,7 +1,7 @@
 function gui(obj)
 %gui Graphical User Interface to mic.stage3D.abstract
 
-h = findall(0,'tag','MIC_3DStage_Abstract_gui');
+h = findall(0,'tag','mic.stage3D.abstract_gui');
 %Prevent opening more than one figure for same instrument
 if ~(isempty(h))
     figure(h);
@@ -25,7 +25,7 @@ PositionUnit=obj.PositionUnit;
 guiFig = figure('Units','pixels','Position',[xst yst xsz ysz],...
     'MenuBar','none','ToolBar','none','Visible','on',...
     'NumberTitle','off','UserData',0,...
-    'Tag','MIC_3DStage_Abstract_gui','HandleVisibility','off');
+    'Tag','mic.stage3D.abstract_gui','HandleVisibility','off');
 defaultBackground = get(0,'defaultUicontrolBackgroundColor');
 set(guiFig,'Color',defaultBackground)
 
