### 2024-10-08 5c72e68
Fix spacing in MICclassesHier.md
