+++ doc/MICclassesHier.md
@@ -7,7 +7,7 @@
 [mic.Joystick](../src/+mic/@Joystick/Readme.md) | control the TIRF stage with a joystick
 [mic.abstract](../src/+mic/@abstract/Readme.md) | Matlab Instrument Control abstract class
 &nbsp;&nbsp;&nbsp;[mic.camera.abstract](../src/+mic/+camera/@abstract/Readme.md) | common interface for all cameras
-&nbsp;&nbsp;&nbsp;[mic.camera.example](../src/+mic/+camera/@example/Readme.md) | template for implementing classes inheriting from mic.camera.abstract
+&nbsp;&nbsp;&nbsp;&nbsp;&bnsp;&bnsp;[mic.camera.example](../src/+mic/+camera/@example/Readme.md) | template for implementing classes inheriting from mic.camera.abstract
 &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[mic.camera.AndorCamera](../src/+mic/+camera/@AndorCamera/Readme.md) | control an Andor camera
 &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[mic.camera.AndorCameraZyla](../src/+mic/+camera/@AndorCameraZyla/Readme.md) | control an Andor camera
 &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[mic.camera.DCAM4Camera](../src/+mic/+camera/@DCAM4Camera/Readme.md) | control a Hamamatsu camera using the DCAM4 API
@@ -19,7 +19,7 @@
 &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[mic.camera.ThorlabsSICamera](../src/+mic/+camera/@ThorlabsSICamera/Readme.md) | control a Thorlabs Scientific Camera via a USB port
 &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[mic.camera.IRSyringPump](../src/+mic/+camera/@IRSyringPump/Readme.md) | use the Syringe Pump at the same time data is taken with IRCamera
 &nbsp;&nbsp;&nbsp;[mic.lightsource.abstract](../src/+mic/+lightsource/@abstract/Readme.md) | abstract class for all light source devices
-&nbsp;&nbsp;&nbsp;[mic.lightsource.example](../src/+mic/+lightsource/@example/Readme.md) | template for implementing classes inheriting from mic.lightsource.abstract
+&nbsp;&bnsp;&bnsp;&nbsp;&nbsp;&nbsp;[mic.lightsource.example](../src/+mic/+lightsource/@example/Readme.md) | template for implementing classes inheriting from mic.lightsource.abstract
 &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[mic.lightsource.CoherentLaser561](../src/+mic/+lightsource/@CoherentLaser561/Readme.md) | coherent Sapphire laser 561
 &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[mic.lightsource.CrystaLaser405](../src/+mic/+lightsource/@CrystaLaser405/Readme.md) | CrystaLaser 405 nm
 &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[mic.lightsource.CrystaLaser561](../src/+mic/+lightsource/@CrystaLaser561/Readme.md) | CrystaLaser 561 nm
@@ -35,15 +35,15 @@
[…]
