### 2024-10-11 a06c594
simulated and funcTest
I changed the names of the Example instruments to simulated instruments. Similarly I renamed the unitTest to funcTest, which is functional testing.
