+++ @simulated_PowerMeter/simulated_PowerMeter.m
@@ -1,5 +1,5 @@
-classdef Example_PowerMeter < MIC_PowerMeter_Abstract
-    % ExamplePowerMeter Class for controlling Example Power Meter.
+classdef simulated_PowerMeter < MIC_PowerMeter_Abstract
+    % SimulatedPowerMeter Class for controlling Example Power Meter.
     % This class provides an interface to the example power meter,
     % implementing all necessary methods to operate the device and manage
     % data acquisition and GUI representation.
@@ -11,10 +11,10 @@ classdef Example_PowerMeter < MIC_PowerMeter_Abstract
     
     % CITATION: Sajjad Khan, Lidkelab, 2024.
     properties (SetAccess = protected)
-        InstrumentName = 'ExamplePowerMeter';
+        InstrumentName = 'SimulatedPowerMeter';
     end
     methods
-        function obj = Example_PowerMeter()
+        function obj = simulated_PowerMeter()
             % Constructor for ExamplePowerMeter
             obj = obj@MIC_PowerMeter_Abstract(~nargout);
             
@@ -273,13 +273,13 @@ classdef Example_PowerMeter < MIC_PowerMeter_Abstract
     end
     
     methods (Static)
-        function Success = unitTest(testMode)
+        function Success = funcTest(testMode)
             if nargin < 1
                 testMode = true;  % Default to test mode
             end
             
             fprintf('Creating instance of ExamplePowerMeter...\n');
-            pm = Example_PowerMeter();  % Create instance
+            pm = simulated_PowerMeter();  % Create instance
             Success = true;
             
             try
