### 2024-10-18 LidkeLab/matlab-instrument-control#142 @bencardoen
Remove IDE files
This comment is associated with a https://github.com/openjournals/joss-reviews/issues/7275.

Including .vscode folders leads to clashes for your users and developers. Update the .gitignore to prevent inclusion.
