### 2024-10-18 LidkeLab/matlab-instrument-control#143 @bencardoen
Abstract Camera code review
This comment is associated with a https://github.com/openjournals/joss-reviews/issues/7275.

- HDF5 does not seem supported?

- Ranges for properties are only partially implemented, yet comments suggest these exist. Please complete these. https://github.com/LidkeLab/matlab-instrument-control/blob/7995050b9c37e0eb8c9ac8a0cbab165b3a9404c9/%40MIC_Camera_Abstract/MIC_Camera_Abstract.m#L51

- A lot of undocumented behavior is hardcoded in an abstract class, this should be documented, and configurable. The code of these methods is on the long side as well, consider refactoring into helper functions. https://github.com/LidkeLab/matlab-instrument-control/blob/main/%40MIC_Camera_Abstract/MIC_Camera_Abstract.m#L134
