### 2024-10-18 LidkeLab/matlab-instrument-control#145 @bencardoen
Documentation improvements
This comment is associated with a https://github.com/openjournals/joss-reviews/issues/7275.

At this point the documentation is extracted from Matlab comments using a custom script.
However, the generated markdown documentation is not 1-1 with the actual code.

There are github actions defined, they do not call the documentation generation script, however.

Please:
- Avoid reinventing the wheel, there are well documented alternatives to generating documentation from matlab annotated source code https://stackoverflow.com/questions/26242145/
- Automate this so documentation is _alway_ 1-1 with code
