+++ src/+mic/@DynamixelServo/DynamixelServo.m
@@ -136,7 +136,7 @@ classdef DynamixelServo < mic.abstract
             end
             obj.Id = Id;          
             if ~exist('Bps','var')
-                Bps = DynamixelServo.DEFAULT_BAUDNUM;
+                Bps = mic.DynamixelServo.DEFAULT_BAUDNUM;
             end
             obj.Bps = Bps;            
             mic.DynamixelServo.loadlibrary();
@@ -169,7 +169,7 @@ classdef DynamixelServo < mic.abstract
         function checkCommStatus(obj)
             % Checks status of COM port
             CommStatus = int32(obj.callDynamixel('dxl_get_result'));
-            if CommStatus == DynamixelServo.COMM_RXSUCCESS
+            if CommStatus == mic.DynamixelServo.COMM_RXSUCCESS
                 mic.DynamixelServo.printErrorCode();
             else
                 mic.DynamixelServo.printCommStatus(CommStatus);
@@ -190,7 +190,7 @@ classdef DynamixelServo < mic.abstract
         
         function value = get.GoalPosition(obj)
             % Gets position
-            value = int32(obj.callDynamixel('dxl_read_word',obj.Id,DynamixelServo.GOAL_POSITION));
+            value = int32(obj.callDynamixel('dxl_read_word',obj.Id,mic.DynamixelServo.GOAL_POSITION));
             %fprintf('get goal %d\n', value);
             obj.checkCommStatus();
         end
@@ -200,14 +200,14 @@ classdef DynamixelServo < mic.abstract
             if value < 0 || value > 1023
                 error('mic.DynamixelServo:PosOutofRange','Position must be between 0 and 1023');
             end
-            obj.callDynamixel('dxl_write_word',obj.Id,DynamixelServo.GOAL_POSITION,value);
+            obj.callDynamixel('dxl_write_word',obj.Id,mic.DynamixelServo.GOAL_POSITION,value);
             %fprintf('set goal %d\n', value);
             obj.checkCommStatus();
         end
         
         function value = get.Led(obj)
             % Gets LED information
-            value = int32(obj.callDynamixel('dxl_read_word',obj.Id,DynamixelServo.LED));
+            value = int32(obj.callDynamixel('dxl_read_word',obj.Id,mic.DynamixelServo.LED));
             obj.checkCommStatus();
         end
         
@@ -216,25 +216,25 @@ classdef DynamixelServo < mic.abstract
             if value ~= 0 && value ~= 1
                 error('mic.DynamixelServo:Led', 'Led can only be 0 off or 1 on');
             end
-            obj.callDynamixel('dxl_write_word',obj.Id,DynamixelServo.LED,value);
+            obj.callDynamixel('dxl_write_word',obj.Id,mic.DynamixelServo.LED,value);
             obj.checkCommStatus();
         end
         
         function value = get.Model(obj)
             % Gets model number
-            value = obj.callDynamixel('dxl_read_word',obj.Id,DynamixelServo.MODEL_NUMBER);
+            value = obj.callDynamixel('dxl_read_word',obj.Id,mic.DynamixelServo.MODEL_NUMBER);
             obj.checkCommStatus();
         end
         
         function value = get.Moving(obj)
             % Gets moving information
-            value = int32(obj.callDynamixel('dxl_read_word',obj.Id,DynamixelServo.MOVING));
+            value = int32(obj.callDynamixel('dxl_read_word',obj.Id,mic.DynamixelServo.MOVING));
             obj.checkCommStatus();
         end
         
         function value = get.MovingSpeed(obj)
             % Gets moving speed
[…]
