### 2024-11-01 636d827
Fix SMP40 and Triggerscope global vars with same name as (new) fun name.
