+++ src/+mic/@Triggerscope/executeCommand.m
@@ -24,7 +24,7 @@ assert(any(contains(Command, obj.CommandList)), ...
     sprintf('Command ''%s'' not listed in obj.CommandList. ', Command));
 
 % Flush the input and output buffers.
-flush(obj.Triggerscope);
+flush(obj.TriggerscopeSerialPortDev);
 
 % Execute the command and check for a response.
 obj.ActivityMessage = sprintf(['Writing command ''%s'' ', ...
