### 2024-11-01 290c7ac
Renamed Example to simulated at top-level src/+mic.
