### 2024-11-10 bef7e14
Address JOSS reviewer comments
Here I addressed comments related to camera/abstract, Attenuator, BioChemValve, GalvoAnalog and funcTests.
