### 2024-11-11 c790111
Update Attenuator file
Added more documentation in Attenuator.m file
