### 2024-11-14 fac5c0c
NanoMaxPiezo properties in md file
