+++ src/+mic/+stage3D/@NanoMaxPiezos/NanoMaxPiezos.m
@@ -19,6 +19,57 @@ classdef NanoMaxPiezos < mic.stage3D.abstract
     % ## Installation Notes
     % Before using this class, ensure that all dependent classes and required Thorlabs drivers are installed and properly configured on your system.
     %
+    % ## Public Properties
+    %
+    % ### `ControllerXSerialNum`
+    % x piezo controller serial number (string).
+    %
+    % ### `ControllerYSerialNum`
+    % y piezo controller serial number (string).
+    %
+    % ### `ControllerZSerialNum`
+    % z piezo controller serial number (string).
+    %
+    % ### `MaxPiezoConnectAttempts`
+    % Maximum attempts to connect to a piezo (default: `1`).
+    %
+    % ### `StrainGaugeXSerialNum`
+    % x piezo strain gauge serial number (string).
+    %
+    % ### `StrainGaugeYSerialNum`
+    % y piezo strain gauge serial number (string).
+    %
+    % ### `StrainGaugeZSerialNum`
+    % z piezo strain gauge serial number (string).
+    %
+    % ### `StagePiezoX`
+    % Piezo object for x position piezo on the stage.
+    %
+    % ### `StagePiezoY`
+    % Piezo object for y position piezo on the stage.
+    %
+    % ### `StagePiezoZ`
+    % Piezo object for z position piezo on the stage.
+    %
+    % ### `StepSize`
+    % Three-element vector specifying the step size in each direction.
+    %
+    % ## Protected Properties
+    %
+    % ### `InstrumentName`
+    % Meaningful instrument name (default: `'NanoMaxStagePiezos'`). This property should not be modified by users.
+    %
+    % ### `Position`
+    % Vector `[x, y, z]` giving the current piezo positions. This property should not be set directly by users.
+    %
+    % ### `PositionUnit`
+    % Units of the position parameter (e.g., `um`, `mm`, etc.). This property should not be set directly by users.
+    %
+    % ## Hidden Properties
+    %
+    % ### `StartGUI`
+    % Flag to control whether the GUI opens on object creation (default: `0`).
+    
     % ## Methods
     % ### `Constructor (mic.stage3D.NanoMaxPiezos())`
     % Initializes piezo controllers for x, y, and z axes based on provided serial numbers. It attempts to connect to the piezos, with error handling to manage connection issues.
