### 2024-11-14 707e996
updated docs of classes
