+++ src/+mic/+lightsource/@ThorlabsLED/ThorlabsLED.m
@@ -19,8 +19,61 @@ classdef ThorlabsLED < mic.lightsource.abstract
     % ## Installation
     % Ensure all required files are in the MATLAB path and that the NI-DAQmx driver is correctly installed and configured on your system.
     %
+    % ## Protected Properties
+    %
+    % ### `InstrumentName`
+    % Name of the instrument.
+    % **Default:** `'ThorlabsLED'`.
+    %
+    % ### `Power`
+    % Currently set output power.
+    % **Default:** `0`.
+    %
+    % ### `PowerUnit`
+    % Unit of power measurement.
+    % **Default:** `'Percent'`.
+    %
+    % ### `MinPower`
+    % Minimum power setting.
+    % **Default:** `0`.
+    %
+    % ### `MaxPower`
+    % Maximum power setting.
+    % **Default:** `100`.
+    %
+    % ### `IsOn`
+    % On or off state of the device (`0` for OFF, `1` for ON).
+    % **Default:** `0`.
+    %
+    % ### `NIDevice`
+    % NIDAQ device name (e.g., `Dev1`).
+    %
+    % ### `AOChannel`
+    % Name of the analog output (AO) channel for the LED on the NIDAQ port (e.g., `ao1`).
+    %
+    % ### `physicalChannel`
+    % Name of the NIDAQ port used for communication.
+    %
+    % ### `V_100`
+    % Voltage at which the current begins to drop from 100%.
+    % **Default:** `5`.
+    %
+    % ### `V_0`
+    % Voltage setting to completely turn off the device.
+    % **Default:** `0`.
+    %
+    % ### `DAQ`
+    % NI DAQ session object.
+    % **Default:** `[]`.
+    %
+    % ## Hidden Properties
+    %
+    % ### `StartGUI`
+    % Indicates whether the GUI should start.
+    % **Default:** `false`.
+    %
     % ## Functions:
-        %   on(obj)
+    %   on(obj)
         %       - Turns the LED lamp ON.
         %       - Usage: obj.on();
         %
