+++ src/+mic/+camera/@ThorlabsIR/ThorlabsIR.m
@@ -11,10 +11,70 @@ classdef ThorlabsIR < mic.camera.abstract
     % For the first time it is required to load the directory of .dll file 
     % from Program Files.    
     %
+    % ## Protected and Transient Properties
+    %
+    % ### `AbortNow`
+    % Flag to stop acquisition.
+    %
+    % ### `FigurePos`
+    % Position of the figure window.
+    %
+    % ### `FigureHandle`
+    % Handle for the figure window.
+    %
+    % ### `ImageHandle`
+    % Handle for the image display.
+    %
+    % ### `ReadyForAcq`
+    % Flag indicating if the camera is ready for acquisition.
+    % **Default:** `0` (call `setup_acquisition` if not ready).
+    %
+    % ### `TextHandle`
+    % Handle for text display.
+    %
+    % ### `dllPath`
+    % Path for the DLL file.
+    %
+    % ## Protected Properties
+    %
+    % ### `CameraIndex`
+    % Index used when more than one camera is present.
+    %
+    % ### `ImageSize`
+    % Size of the current ROI (Region of Interest).
+    %
+    % ### `LastError`
+    % Last error code encountered.
+    %
+    % ### `Manufacturer`
+    % Camera manufacturer.
+    %
+    % ### `Model`
+    % Camera model.
+    %
+    % ### `CameraParameters`
+    % Camera-specific parameters.
+    %
+    % ### `XPixels`
+    % Number of pixels in the first dimension.
+    %
+    % ### `YPixels`
+    % Number of pixels in the second dimension.
+    %
+    % ### `InstrumentName`
+    % Name of the instrument.
+    % **Default:** `'DCxCamera'`.
+    %
+    % ## Hidden Properties
+    %
+    % ### `StartGUI`
+    % Flag to control whether the GUI starts automatically.
+    % **Default:** `false`.
+    
     % ## Constructor
     % obj=mic.camera.ThorlabsIR()
     %
-    % ## Key Function 
+    % ## Key Function
     % initialize, abort, delete, shutdown, getlastimage, getdata, setup_acquisition, start_focus, start_capture, start_sequence, set.ROI, get_Properties, exportState, funcTest  
     %     
     % ## REQUIREMENTS 
