### 2024-11-15 8d4dd8c
Modified createDocumentation script
