+++ src/+mic/+camera/@example/Readme.md
@@ -1,190 +1,191 @@
-# Example_Camera Class Documentation
+# This is an example implementation of mic.camera.abstract
+Matlab Instrument Control Camera Class.
 
-## Class Description
-The `Example_Camera` class is an example implementation of the `MIC_Camera_Abstract` class, simulating a camera with various functionalities.
+REQUIRES:
+mic.camera.abstract.m
 
-### Requirements
-- `MIC_Abstract.m`
+## Protected Properties (Set Access)
 
-## Properties
-### Protected Properties
-- `InstrumentName`: Name of the instrument (`'Simulated Camera'`)
-- `CameraIndex`: Index of the camera (`1`)
-- `ImageSize`: Size of the image (`[1024, 768]`)
-- `LastError`: Last error encountered (initially `''`)
-- `Manufacturer`: Manufacturer of the camera (`'MyCam'`)
-- `Model`: Model of the camera (`'CamModelX100'`)
-- `CameraParameters`: Struct containing camera parameters (`Gain`, `FrameRate`)
-- `XPixels`: Number of pixels in the x-dimension (`1024`)
-- `YPixels`: Number of pixels in the y-dimension (`768`)
+### `InstrumentName`
+Name of the instrument.
+**Default:** `'Simulated Camera'`.
 
-### Public Properties
-- `Binning`: Binning setting (`[1 1]`)
-- `Data`: Captured data (initially `[]`)
-- `ExpTime_Focus`: Exposure time for focus mode (`0.01`)
-- `ExpTime_Capture`: Exposure time for capture mode (`0.02`)
-- `ExpTime_Sequence`: Exposure time for sequence mode (`0.05`)
-- `ROI`: Region of interest (`[1 1024 1 768]`)
-- `SequenceLength`: Length of the sequence (`10`)
-- `SequenceCycleTime`: Cycle time of the sequence (`0.1`)
+### `CameraIndex`
+Index of the camera.
+**Default:** `1`.
 
-### Protected Properties
-- `AbortNow`: Flag to indicate abortion (`false`)
-- `FigurePos`: Position of the figure
-- `FigureHandle`: Handle for the figure
-- `ImageHandle`: Handle for the image
-- `ReadyForAcq`: Flag to indicate readiness for acquisition (`false`)
-- `TextHandle`: Handle for the text
-- `TimerHandle`: Handle for the timer object
+### `ImageSize`
+Size of the image in pixels `[width, height]`.
+**Default:** `[1024, 768]`.
 
-### Hidden Properties
-- `StartGUI`: Boolean indicating if GUI starts automatically (initially `false`)
+### `LastError`
+String storing the last error encountered.
+**Default:** `''` (empty string).
+
+### `Manufacturer`
+Name of the manufacturer.
+**Default:** `'MyCam'`.
+
+### `Model`
+Model name of the camera.
+**Default:** `'CamModelX100'`.
+
+### `CameraParameters`
+Structure containing camera-specific parameters.
+**Default:** `struct('Gain', 1, 'FrameRate', 30)`.
+
+### `XPixels`
+Number of pixels in the first dimension (width).
+**Default:** `1024`.
+
+### `YPixels`
+Number of pixels in the second dimension (height).
+**Default:** `768`.
+
+## Public Properties
+
+### `Binning`
+Binning settings in the format `[binX binY]`.
+**Default:** `[1 1]`.
+
+### `Data`
+Last acquired data.
+**Default:** `[]` (empty array).
+
+### `ExpTime_Focus`
+Exposure time for focus mode.
+**Default:** `0.01`.
+
+### `ExpTime_Capture`
+Exposure time for capture mode.
[…]
