+++ src/+mic/@StepperMotor/Readme.md
@@ -14,8 +14,25 @@ Actuator Type box, from the popup menu in the top select the device
 that you wish to control (it could be either HS NanoMax 300 X Axis,
 HS NanoMax 300 Y Axis or HS NanoMax 300 Z Axis), then OK and Save.
 
+## Properties
+
+### Protected Properties
+
+#### `InstrumentName`
+- **Description:** Name of the instrument.
+- **Default Value:** `'StepperMotor'`
+
+### Public Properties
+
+#### `StartGUI`
+- **Description:** Indicates if the GUI should be started.
+
+#### `SerialN`
+- **Description:** Serial number of the stepper motor controller. This number can be found on the back of the controller.
+- **Example Value:** For SEQ controller, `70850323`.
+
 ## Constructor
-M = MIC_StepperMotor(70850323)
+M = mic.StepperMotor(70850323)
 
 ## Key Functions:
 constructor(), goHome(), getPosition(), getStatus(),
@@ -25,7 +42,7 @@ getJogStep(), closeSBC(), delete(), exportState()
 ## REQUIREMENTS:
 MATLAB 2014 or higher
 Kinesis software from thorlabs
-MIC_Abstract class.
+Abstract class.
 Access to the mexfunctions for this device. (kinesis_SBC_function).
 
 ### CITATION: Mohamadreza Fazel, Lidkelab, 2017.
