### 2024-11-15 54b590b
updated camera/abstract by removing ics SaveType
