+++ src/+mic/+camera/@abstract/abstract.m
@@ -122,36 +122,27 @@ classdef abstract < mic.abstract
                 case 'dipimage'
                     data=dip_image(data,'uint16');
                 case 'matlab'
-                     %already in uint16   
-            end   
-            
-            switch obj.SaveType
-                case 'mat'
-                    eval([obj.AcquisitionType '=data';]);
-                    params=obj.exportparameters();
-                    params.FileName=filename;
-                    save(filename,obj.AcquisitionType,'params');
-                case 'ics'
-                    writeim(obj.Data,filename,'ics');
+                    %already in uint16
             end
             
-%           switch obj.SaveType
-%                case 'mat'
-%                     eval([obj.AcquisitionType ' = data;']);
-%                     params = obj.exportparameters();
-%                     params.FileName = filename;
-%                     save(filename, obj.AcquisitionType, 'params');
-%             end
-
-% OR: if elseif statement?
-
-%             if strcmp(obj.SaveType, 'mat')
-%                 eval([obj.AcquisitionType ' = data;']);
-%                 params = obj.exportparameters();
-%                 params.FileName = filename;
-%                 save(filename, obj.AcquisitionType, 'params');
-%             end
-
+            %             switch obj.SaveType
+            %                 case 'mat'
+            %                     eval([obj.AcquisitionType '=data';]);
+            %                     params=obj.exportparameters();
+            %                     params.FileName=filename;
+            %                     save(filename,obj.AcquisitionType,'params');
+            %                 case 'ics'
+            %                     writeim(obj.Data,filename,'ics');
+            %             end
+            
+            switch obj.SaveType  % preferred, default change for futrure expansion
+                case 'mat'
+                    eval([obj.AcquisitionType ' = data;']);
+                    params = obj.exportparameters();
+                    params.FileName = filename;
+                    save(filename, obj.AcquisitionType, 'params');
+            end
+            
         end
         
         function params = exportParameters(obj)
@@ -259,24 +250,24 @@ classdef abstract < mic.abstract
             end
         end
         
-        function set.SaveType(obj,in)
-            switch in
-                case 'mat'
-                    obj.SaveType='mat';
-                case 'ics'
-                    obj.SaveType='ics';
-                otherwise
-                    warning('SaveType must be mat or ics. Not changed')
-            end
-        end
-        
-%         function set.SaveType(obj, in)
-%             if strcmp(in, 'mat')
-%                 obj.SaveType = 'mat';
-%             else
-%                 warning('SaveType must be "mat". Not changed');
+%         function set.SaveType(obj,in)
+%             switch in
+%                 case 'mat'
+%                     obj.SaveType='mat';
+%                 case 'ics'
+%                     obj.SaveType='ics';
+%                 otherwise
+%                     warning('SaveType must be mat or ics. Not changed')
 %             end
 %         end
+        
+        function set.SaveType(obj, in)
+            if strcmp(in, 'mat')
+                obj.SaveType = 'mat';
+            else
+                warning('SaveType must be "mat". Not changed');
+            end
+        end
         
         function set.ShowLive(obj,in)
             switch in
