+++ README.md
@@ -10,45 +10,60 @@ The MIC software package uses object-oriented programming where a class defines 
 
  The structure of MIC is designed to ensure scalability and easy integration of new hardware.
 
-### MIC_Abstract
+### @abstract
 - Defines basic functions and properties common across all classes.
-  - `MIC_Abstract`
+  - `abstract`
 
-### MIC_LightSource_Abstract
-- MIC_LightSource_Abstract is the base class for all light sources (e.g., lasers), defining common functions and properties.
-  - `MIC_CoherentLaser561`
-  - `MIC_CrystaLaser561`
-  - `MIC_CrystaLaser405`
-  - `MIC_DHOMLaser532`
-  - `MIC_MPB_Laser`
-  - `MIC_RebelStarLED`
-  - `MIC_VortranLaser488`
-  - `MIC_TubeLaserDiode`
-  - `MIC_TIRFLaser488`
-  - `MIC_HSM_Laser488`
-  - `MIC_IX71_Lamp01`
+### lightSource
+- The lightSource is the base class for all light sources (e.g., lasers), defining common functions and properties.
+  - `abstract`
+  - `CoherentLaser561`
+  - `CrystaLaser561`
+  - `CrystaLaser405`
+  - `DHOMLaser532`
+  - `MPBLaser`
+  - `RebelStarLED`
+  - `VortranLaser488`
+  - `TubeLaserDiode`
+  - `TIRFLaser488`
+  - `HSMLaser488`
+  - `IX71Lamp01`
+  - `example`
 
-### MIC_Camera_Abstract
-- MIC_Camera_Abstract is the base class for all camera-related classes, defining common functions and properties.
-  - `MIC_AndorCamera`
-  - `MIC_HamamatsuCamera`
-  - `MIC_IMGSourceCamera`
-  - `MIC_ThorlabsIR`
+### camera
+- The camera is the base class for all camera-related classes, defining common functions and properties.
+  - `abstract`
+  - `AndorCamera`
+  - `AndorCameraZyla`
+  - `DCam4Camera`
+  - `HamamatsuCamera`
+  - `Imaq`
+  - `IMGSourceCamera`
+  - `ThorlabsIR`
+  - `ThorlabsSICamera`
+  - `PyDcam`
+  - `example`
 
-### MIC_LinearStage_Abstract
-- MIC_LinearStage_Abstract is the abstract class for linear stages.
-  - `MIC_MCLMicroDrive`
-  - `MIC_KCubePiezo`
-  - `MIC_TCubePiezo`
+### linearStage
+- The linearStage is the abstract class for linear stages.
+  - `abstract`
+  - `MCLMicroDrive`
+  - `KCubePiezo`
+  - `TCubePiezo`
+  - `example`
 
-### MIC_3DStage_Abstract
-- MIC_3DStage_Abstract is the abstract class for 3D stages.
-  - `MIC_MCLNanoDrive`
-  - `MIC_NanoMaxPiezos`
+### stage3D
+- The stage3D is a class for 3D stages.
+  - `abstract`
+  - `MCLNanoDrive`
+  - `NanoMaxPiezos`
+  - `example`
   
-### MIC_PowerMeter_Abstract
-- MIC_PowerMeter_Abstract creates an interface with the power meter.
-  - `MIC_PM100D`
+### powermeter
+- The powermeter creates an interface with the power meter.
+  - `abstract`
+  - `PM100D`
+  - `example`
 
 This class structure is integral to the functioning and expansion of our imaging capabilities, facilitating easy maintenance and upgrading of the imaging system components.
 
@@ -70,28 +85,28 @@ Example:
 `Children` contain exportState output from children instrument components (if any) called within the parent instrument class. 
 
 
-### Unit test method
[…]
