+++ doc/MICclasses.md
@@ -2,63 +2,71 @@
 
 |**MIC** classes|description|
 -------------|---
-[@MIC_ActiveReg3D](../@MIC_ActiveReg3D/Readme.md) | active 3D registration using camera and stage control via piezoelectric actuators
-[@MIC_H5](../@MIC_H5/Readme.md) | static methods for working with HDF5 files
-[@MIC_Joystick](../@MIC_Joystick/Readme.md) | control the TIRF stage with a joystick
-[@MIC_Abstract](../@MIC_Abstract/Readme.md) | Matlab Instrument Control abstract class
-&nbsp;&nbsp;&nbsp;[@MIC_3DStage_Abstract](../@MIC_3DStage_Abstract/Readme.md) | abstract class for all stages
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[@MIC_MCLNanoDrive](../@MIC_MCLNanoDrive/Readme.md) | control a 3D Peizo stage from Mad City Labs
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[@MIC_NanoMaxPiezos](../@MIC_NanoMaxPiezos/Readme.md) | in a Thorlabs NanoMax stage
-&nbsp;&nbsp;&nbsp;[@MIC_Camera_Abstract](../@MIC_Camera_Abstract/Readme.md) | common interface for all cameras
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[@MIC_AndorCamera](../@MIC_AndorCamera/Readme.md) | control an Andor camera
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[@MIC_AndorCameraZyla](../@MIC_AndorCameraZyla/Readme.md) | control an Andor camera
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[@MIC_DCAM4Camera](../@MIC_DCAM4Camera/Readme.md) | control a Hamamatsu camera using the DCAM4 API
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[@MIC_HamamatsuCamera](../@MIC_HamamatsuCamera/Readme.md) | control a Hamamatsu camera
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[@MIC_Imaq](../@MIC_Imaq/Readme.md) | class designed for camera control using the Image Acquisition Toolbox
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[@MIC_IMGSourceCamera](../@MIC_IMGSourceCamera/Readme.md) | control an ImagingSource camera
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[@MIC_PyDcam](../@MIC_PyDcam/Readme.md) | control a camera through a Python interface
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[@MIC_ThorlabsIR](../@MIC_ThorlabsIR/Readme.md) | control a Thorlabs IR camera
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[@MIC_ThorlabsSICamera](../@MIC_ThorlabsSICamera/Readme.md) | control a Thorlabs Scientific Camera via a USB port
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[@MIC_IRSyringPump](../@MIC_IRSyringPump/Readme.md) | use the Syringe Pump at the same time data is taken with IRCamera
[…]
