### 2024-11-18 f8cdc1e
Update actions to use newer macros.
