+++ .github/workflows/matlab-tests.yml
@@ -19,7 +19,7 @@ jobs:
     steps:
     # Step 1: Checkout the repository
     - name: Checkout repository
-      uses: actions/checkout@v3
+      uses: actions/checkout@v4
 
     # Step 2: Set up MATLAB
     - name: Set up MATLAB
