### 2024-11-18 cb8874c
Fixed some probblems in paper.bib
