+++ paper/paper.bib
@@ -119,7 +119,7 @@
 # SEQ and TIRF papers citations
 
 @article{10.1371/journal.pone.0203291,
-    author = {Pallikkuth, Sandeep AND Martin, Cheyenne AND Farzam, Farzin AND Edwards, Jeremy S. AND Lakin, Matthew R. AND Lidke, Diane S. AND Lidke, Keith A.},
+    author = {Pallikkuth, Sandeep and Martin, Cheyenne and Farzam, Farzin and Edwards, Jeremy S. and Lakin, Matthew R. and Lidke, Diane S. and Lidke, Keith A.},
     title = {Sequential super-resolution imaging using DNA strand displacement},
     journal = {PLOS ONE},
     publisher = {Public Library of Science},
@@ -150,7 +150,7 @@ The epidermal growth factor receptor (EGFR) is a receptor tyrosine kinase with i
 }
 
 @article{Fazel2022,
-    author = {Fazel, Mohamadreza AND Wester, Michael J. AND Schodt, David J. AND Cruz, Sebastian Restrepo AND Strauss, Sebastian AND Schueder, Florian AND Schlichthaerle, Thomas AND Gillette, Jennifer M. AND Lidke, Diane S. AND Rieger, Bernd AND Jungmann, Ralf AND Lidke, Keith A.},
+    author = {Fazel, Mohamadreza and Wester, Michael J. and Schodt, David J. and Cruz, Sebastian Restrepo and Strauss, Sebastian and Schueder, Florian and Schlichthaerle, Thomas and Gillette, Jennifer M. and Lidke, Diane S. and Rieger, Bernd and Jungmann, Ralf and Lidke, Keith A.},
     title = {High-precision estimation of emitter positions using Bayesian grouping of localizations},
     journal = {Nature Communications},
     volume = {13},
@@ -183,7 +183,7 @@ The epidermal growth factor receptor (EGFR) is a receptor tyrosine kinase with i
 # =============================================================================
 # PYME
 @MISC{PYME2020,
-   AUTHOR = {PYME Developer Community},
+   AUTHOR = {{PYME Developer Community}},
    TITLE = {{The PYthon Microscopy Environment}},
    HOWPUBLISHED = {https://www.python-microscopy.org/},
    YEAR = 2020,
