### 2024-11-18 90f4b9c
Deleted all the old top-level @MIC_ and @simulated classes.
