+++ @simulated_LinearStage/simulated_LinearStage.m
@@ -1,285 +1,0 @@
-classdef simulated_LinearStage < MIC_LinearStage_Abstract
-    % This class is an example implementation of MIC_LinearStage_Abstract.
-    % This class simulates a linear stage that can move along one axis.
-    
-    % REQUIRES:
-    % MIC_LinearStage_Abstract.m
-    
-    % Documentation: For detailed documentation check Readme.md file.
-    
-    % CITATION: Sajjad Khan, Lidkelab, 2024.
-    
-    properties (SetAccess = protected)
-        InstrumentName = 'Simulated Linear Stage'; % Name of the instrument
-        PositionUnit = 'mm';                      % Units of position parameter (e.g., mm)
-        CurrentPosition = 0;                      % Current position of the device
-        MinPosition = 0;                          % Lower limit position
-        MaxPosition = 100;                        % Upper limit position
-        Axis = 'X';                               % Stage axis (X, Y, or Z)
-    end
-    
-    properties (Hidden)
-        StartGUI = false;                         % GUI does not start automatically
-    end
-    
-    methods
-        function obj = simulated_LinearStage()
-            % Call superclass constructor
-            obj = obj@MIC_LinearStage_Abstract(~nargout);
-        end
-        
-        function setPosition(obj, position)
-            % Sets the position of the stage to the specified position
-            % position must be within the bounds [MinPosition, MaxPosition]
-            if position < obj.MinPosition || position > obj.MaxPosition
-                error('Position out of bounds');
-            end
-            obj.CurrentPosition = position;
-            disp(['Position set to ', num2str(obj.CurrentPosition), ' ', obj.PositionUnit]);
-            obj.updateGui();
-        end
-        
-        function pos = getPosition(obj)
-            % Returns the current position of the stage
-            pos = obj.CurrentPosition;
-            disp(['Current Position: ', num2str(pos), ' ', obj.PositionUnit]);
-        end
-        
-        function guiFig = gui(obj)
-            % gui: Graphical User Interface to MIC_LinearStage_Abstract
-            % Functionality:
-            %   Move the stage by moving the slider or clicking the jog buttons
-            %   Outside jog buttons (C) are for coarse steps
-            %   Inside jog buttons (F) are for fine steps
-            %   Fine and Coarse step sizes can be specified
-            %   Mouse scroll wheel will move stage when mouse is over slider
-            %   Mouse wheel action can be set to Fine or Coarse with toggle button
-            %   Position to which stage should move can be set in edit box
-            %
-            % Note: Updating gui from higher level class
-            %   To update gui from higher level class the uicontrol objects can be
-            %   accessed via obj.GuiFigure.Children
-            %   To identify the children which need updating they are given tags:
-            %   Slider has tag "positionSlider"
-            %   Set position edit box has tag "positionEdit"
-            
-            %Prevent opening more than one figure for same instrument
-            if ishandle(obj.GuiFigure)
-                guiFig = obj.GuiFigure;
-                figure(obj.GuiFigure);
-                return
-            end
-            
-            %Open figure
-            guiFig = figure('NumberTitle','off','Resize','off','Units','pixels','MenuBar','none',...
-                'ToolBar','none','Visible','on', 'Position',[600 100 450 300]);
-            % Create a property based on GuiFigure
-            obj.GuiFigure = guiFig;
-            obj.GuiFigure.Name = sprintf('%s %s', obj.InstrumentName, obj.Axis);
-            %Prevent closing after a 'close' or 'close all'
[…]
