### 2024-11-19 c51407d
Added startup.m to repo; small change to top README.
