+++ startup.m
@@ -1,0 +1,4 @@
+% Paths tp be added to startup.m for matlab-instrument-control.
+addpath '~/Documents/MATLAB/matlab-instrument-control/src'
+addpath '~/Documents/MATLAB/matlab-instrument-control/mex64'
+
