### 2024-11-25 f3420ab
Added NI-DAQ add-on info in classes
