+++ src/+mic/+powermeter/@PM100D/PM100D.m
@@ -18,6 +18,9 @@ classdef PM100D < mic.powermeter.abstract
     %
     % ## REQUIREMENTS:
     %    NI_DAQ  (VISA and ICP Interfaces) should be installed.
+    %    Data Acquisition Toolbox Support Package for National Instruments
+    %    NI-DAQmx Devices: This add-on can be installed from link:
+    %    https://www.mathworks.com/matlabcentral/fileexchange/45086-data-acquisition-toolbox-support-package-for-national-instruments-ni-daqmx-devices
     %    MATLAB 2021a or higher.
     %    mic.Abstract.m
     %    mic.powermeter.abstract.m
