+++ src/+mic/@Attenuator/Attenuator.m
@@ -23,6 +23,9 @@ classdef Attenuator < mic.abstract
 % - Data Acquisition Toolbox.
 % - MATLAB NI-DAQmx driver installed via the Support Package Installer.
 % - An NI DAQ device.
+% - Data Acquisition Toolbox Support Package for National Instruments
+%   NI-DAQmx Devices: This add-on can be installed from link:
+%   https://www.mathworks.com/matlabcentral/fileexchange/45086-data-acquisition-toolbox-support-package-for-national-instruments-ni-daqmx-devices
 % 
 % ## Class Properties
 % - `Transmission`: The current transmission setting (% of maximum).
