+++ src/+mic/+powermeter/@PM100D/PM100D.m
@@ -1,10 +1,10 @@
-classdef PM100D < mic.powermeter.abstract 
+classdef PM100D < mic.powermeter.abstract
     % mic.powermeter.PM100D: Matlab Instrument class to control power meter PM100D.
     %
     % ## Description
     % Controls power meter PM100D, gets the current power. It can also gets
     % the current temperature. The wavelengtj of the light can also be
-    % set for power measurement (within the sensor limits). The gui shows 
+    % set for power measurement (within the sensor limits). The gui shows
     % a movie of the plot of the
     % measured power where the shown period can be modified. It also shows
     % the current power and the maximum measured power. To run this code
@@ -26,18 +26,18 @@ classdef PM100D < mic.powermeter.abstract
     %    mic.powermeter.abstract.m
     %
     % ### CITATION: Mohamadreza Fazel, Lidkelab, 2017.
-
+    
     properties (SetAccess=protected)
         InstrumentName='PM100D'; %Name of the instrument.
     end
-
-
+    
+    
     methods (Static)
         function obj=PM100D
             %This is the constructor.
             %example PM = mic.powermeter.PM100D
-            obj@mic.powermeter.abstract(~nargout);    
-           
+            obj@mic.powermeter.abstract(~nargout);
+            
             % Find a VISA-USB object.
             vendorinfo = visadevlist;
             s=vendorinfo(1,1).ResourceName;
@@ -52,40 +52,40 @@ classdef PM100D < mic.powermeter.abstract
             %fopen(obj.VisaObj);
             % Measure the limits of the wavelength.
             obj.Limits=minMaxWavelength(obj);
-        end 
-      function funcTest()
-          %testing the class.
-          try
-          TestObj=mic.powermeter.PM100D();
-          fprintf('Constructor is run and an object of the class is made.\n');
-          Limit = minMaxWavelength(TestObj);
-          fprintf('Min wavelength: %d, Max wavelength: %d\n', Limit(1),Limit(2));
-          getWavelength(TestObj);
-          fprintf('The current wavelength is %d nm.\n',TestObj.Lambda);
-          TestObj.Lambda=600;
-          setWavelength(TestObj);
-          fprintf('The wavelength is set to 600 nm.\n');
-          State=exportState(TestObj);
-          fprintf('The exportState function was successfully tested.\n');
-          TestObj.delete();
-          fprintf('The port is closed and the object is deleted.\n');
-          fprintf('The class is successfully tested :)\n')
-          catch E
-             fprintf('Sorry an error has ocured :(\n'); 
-              E
-          end
-      end
+        end
+        function funcTest()
+            %testing the class.
+            try
+                TestObj=mic.powermeter.PM100D();
+                fprintf('Constructor is run and an object of the class is made.\n');
+                Limit = minMaxWavelength(TestObj);
+                fprintf('Min wavelength: %d, Max wavelength: %d\n', Limit(1),Limit(2));
+                getWavelength(TestObj);
+                fprintf('The current wavelength is %d nm.\n',TestObj.Lambda);
+                TestObj.Lambda=600;
+                setWavelength(TestObj);
+                fprintf('The wavelength is set to 600 nm.\n');
+                State=exportState(TestObj);
+                fprintf('The exportState function was successfully tested.\n');
+                TestObj.delete();
+                fprintf('The port is closed and the object is deleted.\n');
+                fprintf('The class is successfully tested :)\n')
+            catch E
+                fprintf('Sorry an error has ocured :(\n');
+                E
+            end
+        end
     end
-
+    
     methods
-
+        
         function  State=exportState(obj)
             %Exporting the class properties to the State.
             State.InstrunetName=obj.InstrumentName;
             State.Lambda=obj.Lambda;
             State.Limits=obj.Limits;
         end
-
+        
[…]
