### 2024-11-26 e1d713e
Update startup.m
