+++ source/MIC/DCAM4/DCAM4UnInit/mexFunction.cpp
@@ -1,16 +1,0 @@
-#include "stdafx.h"
-
-// [] = DCAM4UnInit()
-// Un-initialize the DCAM-API and force close any devices in use.
-void mexFunction(int nlhs, mxArray* plhs[], int	nrhs, const	mxArray* prhs[]) 
-{
-	// Un-initialize all devices.
-	DCAMERR error;
-	error = dcamapi_uninit();
-	if (failed(error))
-	{
-		mexPrintf("Error = 0x%08lX\ndcamapi_uninit() failed.\n", error);
-	}
-
-	return;
-}
