### 2024-11-27 305f963
Added further doc in camera abstract
