+++ startup.m
@@ -1,5 +1,4 @@
 % Paths tp be added to startup.m for matlab-instrument-control.
-addpath '~/Documents/MATLAB/matlab-instrument-control/src'
-addpath '~/Documents/MATLAB/matlab-instrument-control/mex64'
-addpath '~/Documents/MATLAB/matlab-instrument-control/mex64/dcam4'
-
+addpath '~\Documents\MATLAB\matlab-instrument-control\src'
+addpath '~\Documents\MATLAB\matlab-instrument-control\mex64'
+addpath '~\Documents\MATLAB\matlab-instrument-control\mex64\dcam4'
