+++ README.md
@@ -15,7 +15,7 @@ The MIC software package uses object-oriented programming where a class defines 
   - `abstract`
 
 ### camera
-- The camera is the base class for all camera-related classes, defining common functions and properties.
+- The camera namespace is the base for all camera-related classes, defining common functions and properties.
   - `abstract`
   - `AndorCamera`
   - `AndorCameraZyla`
@@ -23,43 +23,43 @@ The MIC software package uses object-oriented programming where a class defines 
   - `HamamatsuCamera`
   - `Imaq`
   - `IMGSourceCamera`
+  - `PyDcam`
   - `ThorlabsIR`
   - `ThorlabsSICamera`
-  - `PyDcam`
   - `example`
 
 ### lightSource
-- The lightSource is the base class for all light sources (e.g., lasers), defining common functions and properties.
+- The lightSource namespace is the base for all light sources (e.g., lasers), defining common functions and properties.
   - `abstract`
   - `CoherentLaser561`
-  - `CrystaLaser561`
   - `CrystaLaser405`
+  - `CrystaLaser561`
   - `DHOMLaser532`
-  - `MPBLaser`
-  - `RebelStarLED`
-  - `VortranLaser488`
-  - `TubeLaserDiode`
-  - `TIRFLaser488`
   - `HSMLaser488`
   - `IX71Lamp01`
+  - `MPBLaser`
+  - `RebelStarLED`
+  - `TIRFLaser488`
+  - `TubeLaserDiode`
+  - `VortranLaser488`
   - `example`
 
 ### linearStage
-- The linearStage is the abstract class for linear stages.
+- The linearStage abstract is the abstract class for linear stages.
   - `abstract`
-  - `MCLMicroDrive`
   - `KCubePiezo`
+  - `MCLMicroDrive`
   - `TCubePiezo`
   - `example`
 
 ### powermeter
-- The powermeter creates an interface with the power meter.
+- The powermeter namespace creates an interface with the power meter.
   - `abstract`
   - `PM100D`
   - `example`
   
 ### stage3D
-- The stage3D is a class for 3D stages.
+- The stage3D is the namespace for 3D stages.
   - `abstract`
   - `MCLNanoDrive`
   - `NanoMaxPiezos`
@@ -69,7 +69,7 @@ This class structure is integral to the functioning and expansion of our imaging
 
 
 ## Common Features
-Each of the instrument components in MIC have constructor, export methods, unit tests, and graphical user interfaces with a common format.
+Each of the instrument components in MIC have constructor, export methods, functional tests, and graphical user interfaces with a common format.
 
 ### Constructor method
 The constructor method of each instrument component class is used to create an object of the class. The constructor method is called when the class is initialized. The constructor method can take input arguments. One example of the constructor method is shown below.
@@ -91,6 +91,7 @@ The current state of the individual instrument can be obtained using the functio
 
 Example: 
 ```
+TIRFLaser488 = mic.lightsource.TIRFLaser488(...)
 [Attributes, Data, Children] = TIRFLaser488.exportState()
 ```
 `Attributes` is a structure with fields carrying information on the current state of the instrument. In the example, Attribute is a structure with fields `Power`, `IsOn` and `InstrumentName`. 
@@ -101,17 +102,18 @@ Example:
 
 
 ### Functional test (funcTest) method
-Each instrument component class in MIC comes equipped with a static method unit test. The `funcTest` function cycles through a series of pre-defined tests, uniquely selected for the corresponding instrument component, outputting success status. Common steps in the functional test method are creating the object, turning the instrument On/Off, change/modify state of the instrument, output exportState and deleting the object.  
[…]
