### 2024-11-27 #7275 @sajjad88
@adamltyson we have a question for @raacampbell. In his comment on "Tidy properties in concrete classes that inherit from an abstract class #133", could you clarify this: "If this is being done to avoid having to define abstract properties to get a concrete class, then why not just do not set the optional properties as Abstract"?
Thank you.
