### 2024-11-28 #7275 @adamltyson
Thanks @sajjad88, I'll let @raacampbell respond, but just FYI the issue is https://github.com/LidkeLab/matlab-instrument-control/issues/133, `#133` refers to another review issue in this repository.
