### 2024-11-28 #7275 @adamltyson
> could you clarify this: "If this is being done to avoid having to define abstract properties to get a concrete class, then why not just do not set the optional properties as Abstract"?

@raacampbell, I think this is the aspect @sajjad88 wanted some clarification on
