### 2024-11-28 #7275 @raacampbell
Was just trying to understand why all the property declaration duplication is happening. I don't have MATLAB handy right now, so I can't confirm, but as I recall if you have non-abstract properties in the abstract class then they need to be  defined in the concrete class for it to be instantiated. If they are abstract then they don't need to be. I think you know this based on your code.
