### 2024-12-03 b1ab455
Added sample mic working example
