+++ paper/paper.bib
@@ -118,19 +118,20 @@
 
 # SEQ and TIRF papers citations
 
-@article{10.1371/journal.pone.0203291,
-    author = {Pallikkuth, Sandeep and Martin, Cheyenne and Farzam, Farzin and Edwards, Jeremy S. and Lakin, Matthew R. and Lidke, Diane S. and Lidke, Keith A.},
-    title = {Sequential super-resolution imaging using DNA strand displacement},
-    journal = {PLOS ONE},
-    publisher = {Public Library of Science},
-    year = {2018},
-    month = {08},
-    volume = {13},
-    number = {8},
-    pages = {1-11},
-    doi = {10.1371/journal.pone.0203291},
-    url = {https://doi.org/10.1371/journal.pone.0203291},
-    abstract = {Sequential labeling and imaging in fluorescence microscopy allows the imaging of multiple structures in the same cell using a single fluorophore species. In super-resolution applications, the optimal dye suited to the method can be chosen, the optical setup can be simpler and there are no chromatic aberrations between images of different structures. We describe a method based on DNA strand displacement that can be used to quickly and easily perform the labeling and removal of the fluorophores during each sequence. Site-specific tags are conjugated with unique and orthogonal single stranded DNA. Labeling for a particular structure is achieved by hybridization of antibody-bound DNA with a complimentary dye-labeled strand. After imaging, the dye is removed using toehold-mediated strand displacement, in which an invader strand competes off the dye-labeled strand than can be subsequently washed away. Labeling and removal of each DNA-species requires only a few minutes. We demonstrate the concept using sequential dSTORM super-resolution for multiplex imaging of subcellular structures.}
+@article{Schodt:23,
+author = {David J. Schodt and Farzin Farzam and Sheng Liu and Keith A. Lidke},
+journal = {Biomed. Opt. Express},
+keywords = {Deformable mirrors; Fluorescence microscopy; High power lasers; Real time imaging; Spatial light modulators; Super resolution microscopy},
+number = {1},
+pages = {429--440},
+publisher = {Optica Publishing Group},
+title = {Automated multi-target super-resolution microscopy with trust regions},
+volume = {14},
+month = {Jan},
+year = {2023},
+url = {https://opg.optica.org/boe/abstract.cfm?URI=boe-14-1-429},
+doi = {10.1364/BOE.477501},
+abstract = {We describe a dedicated microscope for automated sequential localization microscopy which we term Sequential Super-resolution Microscope (SeqSRM). This microscope automates precise stage stabilization on the order of 5-10 nanometers and data acquisition of all user-selected cells on a coverslip, limiting user interaction to only cell selection and buffer exchanges during sequential relabeling. We additionally demonstrate that nanometer-scale changes to cell morphology affect the fidelity of the resulting multi-target super-resolution overlay reconstructions generated by sequential super-resolution microscopy, and that regions affected by these shifts can be reliably detected and masked out using brightfield images collected periodically throughout the experiment. The SeqSRM enables automated multi-target imaging on multiple user-selected cells without the need for multiple distinct fluorophores and emission channels, while ensuring that the resulting multi-target localization data accurately reflect the relative organization of the underlying targets.},
 }
 
 @article{MUDUMBI2024113603,
