+++ paper/paper.md
@@ -59,7 +59,7 @@ Development of new microscopy methods often require integration of novel compone
 
 MIC not only supports the development and control of new instruments but also integrates seamlessly with MATLAB's comprehensive data and image analysis tools, as well as with SMITE[@Schodt:2023] allowing researchers to process data in real-time during experiments. This capability is crucial for iterative testing and development in experimental setups. For example, the SMLM/SPT analysis software suite SMITE was developed in sync with MIC, so acquiring data and performing analyses flow smoothly. Additionally, MIC is designed to be user-friendly for those familiar with MATLAB, offering customizable control classes, extensive export methods, unit tests, and graphical user interfaces for each instrument component.
 
-An example of MIC’s utility is demonstrated through a custom-built Sequential microscope [@10.1371/journal.pone.0203291] specifically designed for dSTORM and DNA-PAINT based super-resolution. Another usage of MIC is demonstrated through a class designed for TIRF based super-resolution imaging systems [@Fazel2022]. Each of these classes includes an intuitive graphical user interface that manages multiple excitation lasers and camera settings, simplifying complex data collection tasks. The MIC GUIs are represented in Figure-1.![MIC GUIs](MIC_overview2.png) *Figure-1: MIC Graphical User Interfaces(GUIs); (a) Camera control GUI (b) Light source (TIRF Laser488) GUI (c) Linear stage (TCubePiezo X) GUI (d) 3D stage (NanoMax Stage Control) GUI (e) Power meter GUI*.
+An example of MIC’s utility is demonstrated through a custom-built Sequential microscope [@Schodt:23] specifically designed for dSTORM and DNA-PAINT based super-resolution. Another usage of MIC is demonstrated through a class designed for TIRF based super-resolution imaging systems [@Fazel2022]. Each of these classes includes an intuitive graphical user interface that manages multiple excitation lasers and camera settings, simplifying complex data collection tasks. The MIC GUIs are represented in Figure-1.![MIC GUIs](MIC_overview2.png) *Figure-1: MIC Graphical User Interfaces(GUIs); (a) Camera control GUI (b) Light source (TIRF Laser488) GUI (c) Linear stage (TCubePiezo X) GUI (d) 3D stage (NanoMax Stage Control) GUI (e) Power meter GUI*.
 
 MIC is designed to operate with HDF5 (Hierarchical Data Format) files which efficiently store very large datasets. The HDF5 format is particularly useful for storing large datasets, as it allows for efficient data storage and retrieval. This is especially important for single molecule fluorescence imaging, where large amounts of data are generated during experiments. The HDF5 format is also supported by MATLAB, which makes it easy to import and analyze data stored in this format. 
 
