### 2024-12-03 d0752a1
Added core testing of simulated classes (and so corresponding abstracts).
