### 2024-12-03 1262e48
updated mic_example.m
