### 2024-12-03 b0a6099
Add mic_example to the main README.
