+++ paper/paper.bib
@@ -97,7 +97,7 @@
 # =============================================================================
 
 # SMITE
-@ARTICLE{Schodt2023,
+@ARTICLE{SchodtWester2023, 
    AUTHOR = {David J. Schodt and Michael J. Wester and Mohamadreza Fazel and
              Sajjad Khan and Hanieh Mazloom-Farsibaf and Sandeep Pallikkuth and
              Marjolein B. m. Meddens and Farzin Farzam and Eric A. Burns and
