+++ paper/paper.bib
@@ -100,7 +100,7 @@
 @ARTICLE{SchodtWester2023, 
    AUTHOR = {David J. Schodt and Michael J. Wester and Mohamadreza Fazel and
              Sajjad Khan and Hanieh Mazloom-Farsibaf and Sandeep Pallikkuth and
-             Marjolein B. m. Meddens and Farzin Farzam and Eric A. Burns and
+             Marjolein B. M. Meddens and Farzin Farzam and Eric A. Burns and
              William K. Kanagy and Derek A. Rinaldi and Elton Jhamba and
              Sheng Liu and Peter K. Relich and Mark J. Olah and
              Stanly L. Steinberg and Keith A. Lidke},
