+++ paper/paper.md
@@ -72,14 +72,14 @@ There are a few other software packages that allow users to control and synchron
 KAL conceived and supervised the development of MIC.
 SK wrote the manuscript.
 SK & MJW added GitHub documentation and converted MIC to a hierarchical organization.
-SK added the example classes with their GUIs, and added unit tests. 
+SK added the example classes with their GUIs, and added functional tests. 
 KAL added AndorCamera and KCubePiezo. 
 SP added AndorCameraZyla and various laser classes. SK tested the laser classes.
 SL added DCAM4Camera to control Hamamatsu cameras using the DCAM4 API.
 MBMM added DynamixelServo, GalvoAnalog, and HamamatsuLCOS class.
 DJS added 3D registration, BiochemValve, CavroSyringePump, H5, and MCLMicroDrive.
 MF added Attenuator and StepperMotor. 
-HMF added 3DStage_Abstract, CoherentLaser561, GalvoDigital, IRSyringPump, and MCLNanoDrive.
+HMF added stage3D.abstract, CoherentLaser561, GalvoDigital, IRSyringPump, and MCLNanoDrive.
 ET added DMP40 and updated Triggerscope.
 FF added FlipMountTTL, ShutterTTL. 
 SP & SK added NanoMax.
