+++ .github/workflows/matlab-tests.yml
@@ -26,6 +26,12 @@ jobs:
       uses: matlab-actions/setup-matlab@v1
 
     # Step 3: Run the MATLAB test script
-    - name: Run MATLAB tests
-      run: |
-        matlab -batch "run('test_core.m');"
+    #- name: Run MATLAB tests
+     # run: |
+      #  matlab -batch "run('runtests.m');"
+
+      # Runs a set of commands using the runners shell
+      - name: Run all tests
+        uses: matlab-actions/run-tests@v1
+        #with:
+         # source-folder: MATLAB
