+++ .github/workflows/matlab-tests.yml
@@ -17,21 +17,21 @@ jobs:
     runs-on: ubuntu-latest
 
     steps:
-    # Step 1: Checkout the repository
-    - name: Checkout repository
-      uses: actions/checkout@v3
+      # Step 1: Checkout the repository
+      - name: Checkout repository
+        uses: actions/checkout@v3
 
-    # Step 2: Set up MATLAB
-    - name: Set up MATLAB
-      uses: matlab-actions/setup-matlab@v1
+      # Step 2: Set up MATLAB
+      - name: Setup MATLAB
+        uses: matlab-actions/setup-matlab@v1
 
-    # Step 3: Run the MATLAB test script
-    #- name: Run MATLAB tests
-     # run: |
+      # Step 3: Run the MATLAB test script
+      #- name: Run MATLAB tests
+      # run: |
       #  matlab -batch "run('runtests.m');"
 
       # Runs a set of commands using the runners shell
       - name: Run all tests
         uses: matlab-actions/run-tests@v1
-        #with:
-         # source-folder: MATLAB
+        with:
+          source-folder: MATLAB
