+++ .github/workflows/matlab-tests.yml
@@ -34,4 +34,4 @@ jobs:
       - name: Run all tests
         uses: matlab-actions/run-tests@v1
         with:
-          source-folder: MATLAB
+          source-folder: src
