+++ run_tests_old.m
@@ -1,126 +1,0 @@
-% Run various tests on the example implementation of the Abstract 
-% classes of Matlab-Instrument-Control (MIC).  
-
-% mic.camera.example
-
-fprintf('mic.camera,example.funcTest ...\n');
-try
-   mic.camera.example.funcTest();
-end
-
-% mic.lightsource.example
-
-fprintf('mic.lightsource.example.funcTest ...\n');
-try
-   mic.lightsource.example.funcTest()
-end
-
-% mic.linearstage.example
-
-fprintf('mic.linearstage.example.funcTest ...\n');
-try
-   mic.linearstage.example.funcTest()
-end
-
-% mic.powermeter.example
-
-fprintf('mic.powermeter.example.funcTest ...\n');
-try
-   mic.powermeter.example.funcTest()
-end
-
-% mic.stage3D.example
-
-fprintf('mic.stage3D.example.funcTest ...\n');
-try
-   mic.stage3D.example.funcTest()
-end
-
-%% A modified version
-% Run various tests on the example implementation of the Abstract classes of Matlab-Instrument-Control (MIC).  
-
-% Q: integration of CI/CD (Travis/CirceCI are free for open source) to ensure your project's integrtiy is retained across commits?
-
-% Initialize a log file to capture test results
-logFile = 'test_log.txt';
-fid = fopen(logFile, 'a'); % Open for appending, creating if necessary
-fprintf(fid, 'Test Log - %s\n', datestr(now));
-fprintf(fid, '=======================\n');
-
-% mic.camera.example
-fprintf('mic.camera.example.funcTest ...\n');
-fprintf(fid, 'Running mic.camera.example.funcTest ...\n');
-try
-    mic.camera.example.funcTest();
-    fprintf(fid, 'Success: mic.camera.example.funcTest passed.\n');
-catch ME
-    warning('Error during mic.camera.example.funcTest: %s', ME.message);
-    fprintf(fid, 'Error during mic.camera.example.funcTest: %s\n', ME.message);
-    fprintf(fid, 'Stack trace:\n%s\n', getReport(ME, 'extended'));
-end
-
-% mic.lightsource.example
-fprintf('mic.lightsource.example.funcTest ...\n');
-fprintf(fid, 'Running mic.lightsource.example.funcTest ...\n');
-try
-    mic.lightsource.example.funcTest();
-    fprintf(fid, 'Success: mic.lightsource.example.funcTest passed.\n');
-catch ME
-    warning('Error during mic.lightsource.example.funcTest: %s', ME.message);
-    fprintf(fid, 'Error during mic.lightsource.example.funcTest: %s\n', ME.message);
-    fprintf(fid, 'Stack trace:\n%s\n', getReport(ME, 'extended'));
-end
-
-% mic.linearstage.example
-fprintf('mic.linearstage.example.funcTest ...\n');
-fprintf(fid, 'Running mic.linearstage.example.funcTest ...\n');
-try
-    mic.linearstage.example.funcTest();
-    fprintf(fid, 'Success: mic.linearstage.example.funcTest passed.\n');
-catch ME
-    warning('Error during mic.linearstage.example.funcTest: %s', ME.message);
-    fprintf(fid, 'Error during mic.linearstage.example.funcTest: %s\n', ME.message);
-    fprintf(fid, 'Stack trace:\n%s\n', getReport(ME, 'extended'));
-end
-
-% mic.powermeter.example
-fprintf('mic.powermeter.example.funcTest ...\n');
-fprintf(fid, 'Running mic.powermeter.example.funcTest ...\n');
-try
-    mic.powermeter.example.funcTest();
[…]
