### 2024-12-04 3f1c464
Added gen doc yaml
