+++ .github/workflows/GenerateDocumentation.yaml
@@ -1,10 +1,10 @@
-name: MATLAB Tests
+name: Generate documentation
 
 # Controls when the action will run.
 on:
   # Triggers the workflow on push or pull request events but only for the main branch
   push:
-    branches: [ main ] 
+    branches: [ code_reorg_clean ] 
   pull_request:
     branches: [ main ]
 
@@ -26,12 +26,12 @@ jobs:
         uses: matlab-actions/setup-matlab@v1
 
       # Step 3: Run the MATLAB test script
-      #- name: Run MATLAB tests
-      # run: |
-      #  matlab -batch "run('runtests.m');"
+      - name: Run MATLAB tests
+        run: |
+          matlab -batch "run('doc/createDocumentation.m');"
 
       # Runs a set of commands using the runners shell
-      - name: Run all tests
-        uses: matlab-actions/run-tests@v1
-        with:
-          source-folder: src
+      # - name: Run all tests
+      #   uses: matlab-actions/run-tests@v1
+      #   with:
+      #     source-folder: src
