### 2024-12-06 e1e96df
genDoc experiment 4.
