+++ .github/workflows/GenerateDocumentation.yml
@@ -31,10 +31,10 @@ jobs:
           matlab -batch "genDoc;"
 
       # Step 4; Commit and push changes
-      #- name: Commit and push changes
-      #  run: |
-      #    git config --global user.name "GitHub Actions Bot"
-      #    git config --global user.email "github-actions[bot]@users.noreply.github.com"
-      #    git add src/+mic/*/Readme.md src/+mic/*/*/Readme.md
-      #    git commit -a -m "Update Readme.md via GitHub Actions" || echo "No changes to commit"
-      #    git push
+      - name: Commit and push changes
+        run: |
+          git config --global user.name "GitHub Actions Bot"
+          git config --global user.email "github-actions[bot]@users.noreply.github.com"
+          git add src/+mic/*/Readme.md src/+mic/*/*/Readme.md
+          git commit -a -m "Update Readme.md via GitHub Actions" || echo "No changes to commit"
+          git push
