+++ src/+mic/@H5/Readme.md
@@ -1,6 +1,6 @@
 # mic.H5 Class: mic.H5 A collection of static methods for working with HDF5 files
 
-## Description q
+## Description
 The `mic.H5` class is designed for handling HDF5 file operations in MATLAB.
 It includes static methods to create files, write data asynchronously, and manage data groups within HDF5 files.
 This class leverages MATLAB's capabilities to interact with large datasets efficiently.
