### 2024-12-06 c385b5c
Update CoherentLaser561.m
