### 2024-12-06 52d8b77
Update GenerateDocumentation.yml
