+++ .github/workflows/GenerateDocumentation.yml
@@ -4,7 +4,7 @@ name: Generate documentation
 on:
   # Triggers the workflow on push or pull request events but only for the main branch
   push:
-    branches: [ code_reorg_clean ] 
+    branches: [ main ] 
   pull_request:
     branches: [ main ]
 
