+++ .github/workflows/draft-pdf.yml
@@ -1,4 +1,6 @@
-on: [push]
+on: 
+  push:
+    branches: [ code_reorg_clean ] 
 
 jobs:
   paper:
