### 2024-12-09 LidkeLab/matlab-instrument-control#124 @sajjad88
The hierarchical code/+mic/ is a great idea. Now our new MIC organization is as per the suggestions where we have used src/ as the Matlab source directory. We are providing an example startup.m file in the top level directory for this new code organization.
