### 2024-12-09 LidkeLab/matlab-instrument-control#127 @sajjad88
In the top-level `Readme.md`, we have added two sections. Please see ‘High-level implementation of MIC classes’ and ‘Projects using MIC’.
