### 2024-12-09 LidkeLab/matlab-instrument-control#128 @sajjad88
The .DS store files have been removed and the file specification added to gitignore.
