### 2024-12-09 LidkeLab/matlab-instrument-control#129 @sajjad88
We have added more documentation and expanded the top-level Readme.md
