### 2024-12-09 LidkeLab/matlab-instrument-control#130 @sajjad88
The Example files have been renamed to simulated files.
