### 2024-12-09 LidkeLab/matlab-instrument-control#131 @sajjad88
The clc has been removed and was the only one present in the classes.
