### 2024-12-09 LidkeLab/matlab-instrument-control#133 @sajjad88
Please look at:
https://www.mathworks.com/help/matlab/matlab_oop/using-heterogeneous-class-hierarchies.html
Here, the property “Type” is an abstract property declared in the abstract class with “Abstract, SetAccess = private” (basically a signature or header), while in the subclasses “Type” is declared with “SetAccess = private” only (no “Abstract”) and set to a default value for each subclass. We merely copied this format in our development (and we did it in +linearstage, +lightsource, etc. as well as +camera, but it is most noticeable in +camera). In the subclasses, the properties are assigned values in their declarations, which Abstract (“= True”) implies (true – The property has no implementation, but a concrete subclass must override this property without Abstract set to
true --- see: https://www.mathworks.com/help/matlab/matlab_oop/property-attributes.html). We investigated removing the abstract properties altogether and just make them concrete subclass properties, but this didn’t seem as clean to us in forcing the developer to declare certain local versions of already named abstract properties.
