### 2024-12-09 LidkeLab/matlab-instrument-control#134 @sajjad88
The GalvoAnalog class is designed to set the galvo mirror to a single fixed position corresponding to the applied voltage. It is not intended for dynamic or continuous scanning, for that purpose, we have GalvoDigital class. We have made this more explicit in the corresponding Readme.
