### 2024-12-09 LidkeLab/matlab-instrument-control#142 @sajjad88
We have deleted .vscode folders and added this name to .gitignore to prevent inclusion.
