### 2024-12-09 LidkeLab/matlab-instrument-control#143 @sajjad88
• For your comment about “HDF5 does not seem supported”, we decided to simplify the SaveType to only ‘mat’. However, we can save the data in HDF5 in software related to particular microscopes. For reference, please see ‘High-level implementation of MIC classes’ and ‘Projects using MIC’ in the top-level Readme, where we use higher level data collection implementations build on software in MIC.  We decided not to bundle these higher-level data collection routines with MIC, but separate them into their own repositories depending on microscope type (e.g., TIRF, sequential), and use these software to store results into HDF5.
• We removed the SaveType = ‘ics’, but retained the variable SaveType for future expansion.
• Further documentation has been added to camera abstract class.
