### 2024-12-09 LidkeLab/matlab-instrument-control#145 @sajjad88
We would really like to create .md files for the MIC documentation as they show up nicely when we web browse GitHub (as we did with our previously published JOSS code, SMITE).  MATLAB's "publish" outputs a variety of formats, but not Markdown, and the other tools referenced in the stackoverflow link produce HTML but not Markdown.  GitHub does not natively interpret the file
README.html, so each solution requires some non-standard tweaking.  Given that, we prefer to use our script to generate README.md files directly. We have added a GitHub action to automatically generate the documentation README.md files every time a push or pull request to the main branch is made.
