### 2024-12-09 #7275 @sajjad88
@adamltyson we have a source code in a modified hierarchy. We have a release v0.1.2 which is the newest version of mic as proposed by the reviewers.

@raacampbell and @bencardoen we have addressed your comments in the opened issues.

Thank you to all for your helpful comments/suggestions to improve mic for the better!
