### 2024-12-16 LidkeLab/matlab-instrument-control#124 @raacampbell
The "example" `startup.m` file is not ideal. If the user starts MATLAB in the same directory as the file then it generates errors due to non-existent folders.

```
> In path (line 109)
In addpath>doPathAddition (line 126)
In addpath (line 90)
In startup (line 2)
Warning: Name is nonexistent or not a directory: /home/rob/Downloads/matlab-instrument-control/\Users\sajja\Documents\MATLAB\matlab-instrument-control\src
> In path (line 109)
In addpath>doPathAddition (line 126)
In addpath (line 90)
In startup (line 2)
Warning: Name is nonexistent or not a directory: /home/rob/Downloads/matlab-instrument-control/C
> In path (line 109)
In addpath>doPathAddition (line 126)

etc
```

It is trivial to add these directories to the path so you don't really need an example:
1. Add instructions stating the user should add the three folders to the path and remove the `startup.m`
2. Perhaps better yet, write a function that automatically adds the right folders to the path. The easiest way to do this is to (a) move mex64 into src then (b) you can generate the path to add with `genpath('src')`. [This is what is happening here](https://github.com/SWC-Advanced-Microscopy/BakingTray/blob/master/addBTtopath.m)
