### 2024-12-16 LidkeLab/matlab-instrument-control#125 @raacampbell
I have to think about this but there is for sure one thing to do now: I see that that test class you wrote does not run as expected.

* MATLAB has a command called `runtests` which your test class name clashes with. You class does not run, therefore, if you run `runtests`. Pull request filed.
* The test class should also not be in the main code directory. I have moved it into a directory called tests.  Pull request filed.
* `simulated_tests/test_simulated_Instrument` fails by verification. Please investigate.
