### 2024-12-16 LidkeLab/matlab-instrument-control#134 @raacampbell
Sorry, I don't follow the idea behind using the digital lines to control an analog output signal. I have never seen this before: I think it merits a longer explanation. Is this class related a particular microscope which has an unusual  way of handling galvo scanning? You state "HSM" microscope in the help text of `GalvoDigital` , but I don't know what that is.
