### 2024-12-16 #7275 @raacampbell
I went over the issues. Some I closed. Some I added more comments to as there are outstanding points. Some I have not got to yet. However, there is stuff for the authors to look into now.
