+++ src/+mic/@Attenuator/Attenuator.m
@@ -169,11 +169,13 @@ classdef Attenuator < mic.abstract
         
         function setTransmission(obj,Transmission_in)
             % Sets output attenuation in percentage of maximum
-            if Transmission_in >obj.MaxTransmission+0.001 % to introduce a margin of tolerance for comparisons
-               error('The input transmission is too large, setting to %d\n',obj.MaxTransmission); 
+            if Transmission_in > obj.MaxTransmission
+               warning('The input transmission is too large, setting to %d\n',obj.MaxTransmission); 
+               return; % return statement serves to terminate the function early when specific conditions are met
             end
-            if Transmission_in<obj.MinTransmission-0.001
-               error('The input transmission is too small, setting to %d\n',obj.MinTransmission); 
+            if Transmission_in < obj.MinTransmission
+               warning('The input transmission is too small, setting to %d\n',obj.MinTransmission); 
+               return
             end
             if isempty(obj.Input_Voltage)
                error('The suitable calibration file should be loaded. Please call loadCalibration().'); 
