### 2024-12-16 c53b20b
Added link to LSMAQ in Related Software section of main readme file
