### 2024-12-16 ed1966f
Make setupMIC.m for adding MIC paths in startup.m; some doc fixes for testing
