+++ startup.m
@@ -1,4 +1,0 @@
-% Paths tp be added to startup.m for matlab-instrument-control.
-addpath C:\Users\sajja\Documents\MATLAB\matlab-instrument-control\src
-addpath C:\Users\sajja\Documents\MATLAB\matlab-instrument-control\mex64
-addpath C:\Users\sajja\Documents\MATLAB\matlab-instrument-control\mex64\dcam4
