### 2024-12-16 36f9d1c
More README cleaning.
