+++ README.md
@@ -130,11 +130,11 @@ Each instrument will be controlled by its own drivers, which must be installed o
 
 Similarly, the `MCLMicroDrive` class controls the Mad City Labs Micro Stage and requires the `MicroDrive.dll` dynamic-link library. The first time this class is used on a given computer, the user will be prompted to select the location of `MicroDrive.dll`. On a Windows machine, this is typically placed by default in `C:\Program Files\Mad City Labs\MicroDrive\MicroDrive.dll` during the installation process (installation files provided by MCL).
 
-To set up the correct MATLAB path in startup.m (typically located in the Documents\MATLAB folder of the user's home directory), add the line
+To set up the correct MATLAB path for MIC in startup.m (typically located in the Documents\MATLAB folder in the user's home directory), add the line
 ```matlab
 run(fullfile(userpath, 'matlab-instrument-control', 'setupMIC'))
 ```
-(for MATLAB 2017a or more recent versions)
+(for MATLAB 2017a or more recent versions).
 
 ## Testing
 See [tests](tests) folder for unit tests of simulated classes, which in turn tests the various abstract classes the
