+++ .github/workflows/matlab-tests.yml
@@ -4,7 +4,7 @@ name: MATLAB Tests
 on:
   # Triggers the workflow on push or pull request events but only for the main branch
   push:
-    branches: [ JOSSrev1changes ]
+    branches: [ main ]
   pull_request:
     branches: [ main ]
 
