### 2024-12-17 LidkeLab/matlab-instrument-control#134 @sajjad88
The galvo on Hyper spectral line-scanning microscope (HSM) is not directly controlled by analog input voltage. We used a National Instrument card to send a 16-bit digital signal to the galvo controller. The galvo controller then converts the 16-bit digital signal to an analog voltage that drives the galvo mirror on HSM. See the attached png.
![galvo](https://github.com/user-attachments/assets/6ab8e2ce-1407-4133-8433-4c5a5cef2e3c)
