### 2024-12-17 LidkeLab/matlab-instrument-control#125 @sajjad88
Thanks for the pull requests.  This is definitely a clean way to set the tests up and we have adopted your suggestions.

simulated_tests/test_simulated_Instrument fails by verification.

Running it manually from the tests subdirectory or on a git push in Github actions produces a successful test for us.  This particular test is different from the others as it tries to produce a GUI when run directly, but running it interactively or batch worked just fine when we tried it after the current round of changes have been made.
