### 2024-12-17 LidkeLab/matlab-instrument-control#124 @sajjad88
We added the function setupMIC.m in the top-level directory to set up the proper paths for the matlab-instrument-control repository.  In the README, we added the comment:

To set up the correct MATLAB path for MIC in startup.m (typically located in the Documents\MATLAB folder in the user's home directory), add the line

run(fullfile(userpath, 'matlab-instrument-control', 'setupMIC'))

(for MATLAB 2017a or more recent versions).
