### 2024-12-17 LidkeLab/matlab-instrument-control#126 @sajjad88
We have added LSMAQ in the paper and Readme.md files.
