### 2024-12-17 819d2ab
Update description of GalvoDigital
