### 2024-12-17 b9575cd
Update GalvoDigital.m
