### 2024-12-18 LidkeLab/matlab-instrument-control#125 @raacampbell
Ah, it was the GUI. I was running in `-nojvm` because of a MATLAB bug impacting my machine that TMW can't fix so the GUI was not coming up.
