### 2024-12-18 LidkeLab/matlab-instrument-control#134 @raacampbell
OK, I am being a bit slow here because it seems you are using a galvo controller that I have never encountered before. I initially thought there was something special about the HSM that required the digital word scanner control.

Let me see if I understand correctly: are you saying this Cambridge Technologies galvo controller (what model is it? The README should say but does not) is digital and does not accept an analog waveform at all? Therefore you must generate a sequence of 16 bit digital words in order to generate a changing voltage waveform for scanning? My confusion stems from the fact that other digital controllers I have used, such as the [Saturn 5](https://scannermax.com/products/saturn-5-system), _do_ accept an analog input as a control signal.
