### 2024-12-18 LidkeLab/matlab-instrument-control#134 @sajjad88
The model of galvo mirror on HSM is Cambridge Technology 6210H. The control board (Cambridge Technology 67121HHJ-1 with digital input module 6757) was calibrated by the vendor for this specific scanning mirror such that it controls the mirror movement between -15 and +15 degrees from the center position with a linear response to analog voltage from -10 to 10 Volts. The NI-DAQ card (PCIe-6343) sends a 16-bit binary signal to the control board to set the mirror angle. This galvo mirror can be also controlled with an analogy servo driver (e.g. Cambridge Technology 673XX). However, the NI-DAQ card we used can only output an analog voltage with an Absolute Accuracy at Full Scale of 3 mV, which is larger than the required scanning step of 1 mV for our application. Therefore, we selected the digital servo driver for this application.
