### 2024-12-20 LidkeLab/matlab-instrument-control#134 @raacampbell
It would be helpful to explain this in the README for this class in this sort of detail or higher, as this is unusual.

Largely out of curiosity: why do you need such high absolute accuracy? e.g. even if you are scanning 2048 x 2048 pixels at full scale wouldn't the pointing error be less than 1/3 of a pixel (because it's an absolute error not a relative error)? Since: 20/3300E-6 = 6060
