### 2025-01-03 LidkeLab/matlab-instrument-control#134 @sajjad88
@raacampbell Definitely we'll have these details in the README of the class.

FYI, the scanning step size is determined by the magnification of the imaging system, from our optical design, moving the beam by 100 nm at the sample plane requires galvo to rotate by 0.00184 degree, which corresponds to applying 1mV if using an analog driver.
