### 2025-01-08 LidkeLab/matlab-instrument-control#134 @raacampbell
This is useful information. Can you add to the README, please?
