### 2025-01-08 84584ac
Update GalvoDigital.m
