### 2025-01-08 6ea5376
Update Readme of GalvoDigital
