### 2025-01-08 688aae1
Update GenerateDocumentation.yml
