### 2025-01-10 #7275 @adamltyson
Hi @sajjad88, the paper reads nicely. My only suggestion would be to define some of the abbreviations. Even those familiar with microscopy may not know what e.g. "SMLM/SPT" means.
