+++ paper/paper.md
@@ -40,11 +40,11 @@ authors:
     affiliation: 1
     corresponding: true
 affiliations:
- - name: Department of Physics and Astronomy, University of New Mexico
+ - name: Department of Physics and Astronomy, University of New Mexico, Albuquerque, New Mexico, USA
    index: 1
- - name: Nanoscience and Microsystems Engineering, University of New Mexico
+ - name: Nanoscience and Microsystems Engineering, University of New Mexico, Albuquerque, New Mexico, USA
    index: 2
- - name: Department of Mathematics and Statistics, University of New Mexico
+ - name: Department of Mathematics and Statistics, University of New Mexico, Albuquerque, New Mexico, USA
    index: 3
 date: July 2024
 bibliography: paper.bib
@@ -57,24 +57,24 @@ MATLAB Instrument Control (MIC) is a software package designed to facilitate dat
 # Statement of need
 Development of new microscopy methods often require integration of novel components and/or novel workflows between low-level hardware. Creating such software solutions for each new instrument or for prototyping can significantly slow down the development cycle.  It is also often desirable to be able to quickly analyze incoming data during development.  To address this critical need, we have developed the MATLAB Instrument Control (MIC) software package, freely available and specifically tailored for the customization and automation of complex, multi-component microscopy systems [@Pallikkuth_article:2018]. MIC leverages MATLAB's (MathWorks Inc.) robust environment for object-oriented programming, allowing users to control diverse instrumentation through a unified interface. Each instrument class within MIC inherits from a common abstract class, ensuring consistency while enabling flexibility to accommodate a variety of components like lasers, stages, and cameras. The source code for MIC has been archived to GitHub: [https://github.com/LidkeLab/matlab-instrument-control](https://github.com/LidkeLab/matlab-instrument-control).
 
-MIC not only supports the development and control of new instruments but also integrates seamlessly with MATLAB's comprehensive data and image analysis tools, as well as with SMITE [@SchodtWester2023] allowing researchers to process data in real-time during experiments. This capability is crucial for iterative testing and development in experimental setups. For example, the SMLM/SPT analysis software suite SMITE was developed in sync with MIC, so acquiring data and performing analyses flow smoothly. Additionally, MIC is designed to be user-friendly for those familiar with MATLAB, offering customizable control classes, extensive export methods, functional tests, and graphical user interfaces for each instrument component.
+MIC not only supports the development and control of new instruments but also integrates seamlessly with MATLAB's comprehensive data and image analysis tools, as well as with Single Molecule Imaging Toolbox Extraordinaire (SMITE) [@SchodtWester2023] allowing researchers to process data in real-time during experiments. This capability is crucial for iterative testing and development in experimental setups. For example, the single molecule localization microscopy/single particle tracking (SMLM/SPT) analysis software suite SMITE was developed in sync with MIC, so acquiring data and performing analyses flow smoothly. Additionally, MIC is designed to be user-friendly for those familiar with MATLAB, offering customizable control classes, extensive export methods, functional tests, and graphical user interfaces for each instrument component.
 
[…]
