### 2025-01-10 #7275 @sajjad88
Hello @adamltyson we have modified the paper as per your suggestion. Please review it.

Thank you to both @raacampbell and @bencardoen for taking the time to review the MIC code and provide valuable feedback! Your insights and suggestions really improved the quality of our work significantly.
