### 2025-01-16 #7275 @adamltyson
Hi @sajjad88,

I just need @bencardoen to confirm that they're happy to accept the paper (I assume this is the case), then we can move forwards (@raacampbell has confirmed to me privately).
