### 2025-01-16 34c08ae
updated doi and release version
