### 2025-01-17 #7275 @bencardoen
@adamltyson @sajjad88 yes, I'm happy with the paper moving forward, apologies for the delay, and congrats @sajjad88 & team for an excellent contribution
