### 2025-01-17 #7275 @raacampbell
Yes, very nice. I plan to use your power meter control classes.
