### 2025-01-17 9799dbb
Update CITATION.cff
