+++ CITATION.cff
@@ -1,4 +1,4 @@
-cff-version: 0.9.9
+cff-version: 1.0.0
 message: If you use this software, please cite it as below.
 references:
   - type: software
