### 2025-01-17 609a5c9
Update CITATION.cff
