+++ CITATION.cff
@@ -49,8 +49,8 @@ references:
         affiliation: "Department of Physics and Astronomy, University of New Mexico, Albuquerque, New Mexico, USA"
   title: "A MATLAB-based Instrument Control (MIC) package for fluorescence imaging"
   version: 1.0.0
-  doi: 10.5281/zenodo.14675632
-  date-released: 2025-01-16
+  doi: 
+  date-released: 2025-01-17
   keywords:
     - "MATLAB"
     - "instrument control"
