### 2025-01-17 f3000ae
Update CITATION.cff
