### 2025-01-17 #7275 @sajjad88
Hi @adamltyson we have a new release with version v1.0.0 and the Zenodo DOI is [10.5281/zenodo.14680987](https://doi.org/10.5281/zenodo.14680987). I confirm other tasks are complete. Thank you.
