### 2025-01-20 #7275 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
