### 2025-01-22 #7275 @adamltyson
Hi @sajjad88, the final acceptance decision is down to the Associate Editor-in-Chief, so we'll need to wait until they have some time to look at this submission (this is why the editorialbot commands don't work for you!).
