### 2025-01-28 #7275 @Kevin-Mattheus-Moerman
@sajjad88 et al., as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Below are some final checks.
